checkout test: enable test with complex relative path
This test was added, commented out, in fed1b5ca (git-checkout: Test
for relative path use, 2007-11-09).  Later git's path handling was
improved (d089ebaa, setup: sanitize absolute and funny paths in
get_pathspec(), 2008-01-28) but we forgot to enable the now-working
test.

This test expects to run from a subdirectory, so add a 'cd'.  While
we're here, examine the content of the checked-out file instead of
just checking that it exists.  The other checkout tests already do the
same.

Signed-off-by: Stefan Beller <stefanbeller@googlemail.com>
Signed-off-by: Jonathan Nieder <jrnieder@gmail.com>

Former-commit-id: 1aca67a43ed6e4c002b571313d5cb5c2fe06695b [formerly 1aca67a43ed6e4c002b571313d5cb5c2fe06695b [formerly b0afc02649285febd6155c1c64d2383b856c1c0e]]
Former-commit-id: 81fedb73845110301cb5b3fd1a8fa3ead0457b19
Former-commit-id: 258909cbe7ff4e2b92bfb0bda5ab67972ac6f161