git-prompt.sh: optionally show upstream branch name
When working with multiple remotes, it is common to switch the upstream
from a remote to another. Doing so, the prompt may not be the expected
one. Providing an option to display tracking information sounds useful.

Add a "name" option to GIT_PS1_SHOWUPSTREAM which will show the upstream
abbrev name. This option is ignored if "verbose" is false.

Signed-off-by: Julien Carsique <julien.carsique@gmail.com>
Improved-by: SZEDER Gábor <szeder@ira.uka.de>
Signed-off-by: Jonathan Nieder <jrnieder@gmail.com>

Former-commit-id: 11228d35b7893cdbeeb9e00573729d4be30770ac [formerly 3dffa7c958ee0a91740561830c17e0fdd67c3876] [formerly 3dffa7c958ee0a91740561830c17e0fdd67c3876 [formerly 1f6806cf2da84e30d162ca3aee15d16db4f1daef]]
Former-commit-id: ee4b7f7f790590ed80e8f9a1a5cf05148661369c
Former-commit-id: 8f9dbd06989f15bda73ce5ad739745641c1ca51c