cherry-pick: handle "-" after parsing options
Currently, we only try converting argv[1] from "-" into "@{-1}".  This
means we do not notice "-" when used together with an option.  Worse,
when "git cherry-pick" is run with no options, we segfault.  Fix this
by doing the substitution after we have checked that there is
something in argv to cherry-pick and know any remaining options are
meant for the revision-listing machinery.

This still does not handle "-" after the first non-cherry-pick option.
For example,

	git cherry-pick foo~2 - bar~5

and

	git cherry-pick --no-merges -

will still dump usage.

Reported-by: Stefan Beller <stefanbeller@googlemail.com>
Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Jonathan Nieder <jrnieder@gmail.com>

Former-commit-id: 784210e6bd09485c04efd7ff8a3247a9cdbf86df [formerly 784210e6bd09485c04efd7ff8a3247a9cdbf86df [formerly d644c5502fc576904a07fab35f43540fa9f2e7c2]]
Former-commit-id: f0c651b2a11421ae0f418cfd9d811bc901a5d331
Former-commit-id: 010575927c941d912df2ab5b9a32343d6c122419