path-utils test: rename mingw_path function to print_path
mingw_path was introduced in abd4284 to output a mangled path as it is
passed as an argument to main(). But the name is misleading because
mangling does not come from MinGW, but from MSYS [1]. As abd4284 does not
introduce any MSYS or MinGW specific code but just prints out argv[2] as
it is passed to main(), give the function the more generic and less
confusing name "print_path".

[1] http://www.mingw.org/wiki/Posix_path_conversion

Signed-off-by: Sebastian Schuberth <sschuberth@gmail.com>
Signed-off-by: Jonathan Nieder <jrnieder@gmail.com>

Former-commit-id: 10e61a03756f0cb9adb8a92e585a9eab7daa090d [formerly 60b413ee1249a50730cf4982391164713db6ad6c] [formerly 60b413ee1249a50730cf4982391164713db6ad6c [formerly 7ffd18fce1eef2670661c2e7fd0cb3ef0b8b7fb7]]
Former-commit-id: 8cc9becfd945bad6db0c99dca7fe869ba2f91d01
Former-commit-id: 167f6db1d9baafaa9bfe59e90489f8be884b94a4