git-svn: Warn about changing default for --prefix in Git v2.0
In Git v2.0, we will change the default --prefix for init/clone from
none/empty to "origin/" (which causes SVN-tracking branches to be
placed at refs/remotes/origin/* instead of refs/remotes/*).

This patch warns users about the upcoming change, both in the git-svn
manual page, and on stderr when running init/clone in the "multi-mode"
without providing a --prefix.

Cc: Eric Wong <normalperson@yhbt.net>
Signed-off-by: Johan Herland <johan@herland.net>
Signed-off-by: Eric Wong <normalperson@yhbt.net>

Former-commit-id: 376c8df510bf99a89f06913bd98dbd43f19a8a25 [formerly 376c8df510bf99a89f06913bd98dbd43f19a8a25 [formerly f849bb6b3b4ccc7cc1a68f49d6ff1e9c508fdf17]]
Former-commit-id: 860601d527d3590bbe6a79faf69ec4c19dea1458
Former-commit-id: 26843895e8b986772e25f6b7b9d7d7800a66e58a