test: use unambigous leading path (/foo) for MSYS
In test cases for relative_path, path with one leading character
(such as /a, /x) may be recogonized as "a:/" or "x:/" if there is
such DOS drive on MSYS platform. Use an umambigous leading path
"/foo" instead.

Also change two leading slashes (//) to three leading slashes (///),
otherwize it will be recognized as UNC name on MSYS platform.

Signed-off-by: Jiang Xin <worldhello.net@gmail.com>
Acked-by: Sebastian Schuberth <sschuberth@gmail.com>
Signed-off-by: Jonathan Nieder <jrnieder@gmail.com>

Former-commit-id: aa20c1606a6a24082fbd2dbd1762d3ad68f2f56e [formerly aa20c1606a6a24082fbd2dbd1762d3ad68f2f56e [formerly daf19a80fa6b7402f8a2efa8f8255b4b5d6b48a9]]
Former-commit-id: 9167eb00ba0b031298f77a3ce65d25236d967d8d
Former-commit-id: d2914a4d75b6a5b5b32ba0f7e8b02569f8a96efe