graph: fix coloring around octopus merges
When drawing the graph of an octopus merge, we draw a horizontal line
from parents 3 and above into the asterisk representing the commit. The
sections of this line should be colored to match the graph lines coming
in from above.

However, if the commit is not in the left-most column we do not take
into account the columns to the left of the commit when calculating
these colors. Fix this by adding the appropriate offset to the column
index used for calculating the color.

Signed-off-by: Hemmo Nieminen <hemmo.nieminen@iki.fi>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: fbcc393215f8526c0732ac6f2337f6c597e72087 [formerly fbcc393215f8526c0732ac6f2337f6c597e72087 [formerly 339c17bc7690b5436ac61c996cede3d52c85b50d]]
Former-commit-id: 9888ea1d844c168dfcf3dfa4c432cdc922a56297
Former-commit-id: bfd31a727ddc66bab8e772751fadf8ea17264d1a