C: have space around && and || operators
Correct all hits from

    git grep -e '\(&&\|||\)[^ ]' -e '[^	 ]\(&&\|||\)' -- '*.c'

i.e. && or || operators that are followed by anything but a SP,
or that follow something other than a SP or a HT, so that these
operators have a SP around it when necessary.

We usually refrain from making this kind of a tree-wide change in
order to avoid unnecessary conflicts with other "real work" patches,
but in this case, the end result does not have a potentially
cumbersome tree-wide impact, while this is a tree-wide cleanup.

Fixes to compat/regex/regcomp.c and xdiff/xemit.c are to replace a
HT immediately after && with a SP.

This is based on Felipe's patch to bultin/symbolic-ref.c; I did all
the finding out what other files in the whole tree need to be fixed
and did the fix and also the log message while reviewing that single
liner, so any screw-ups in this version are mine.

Signed-off-by: Felipe Contreras <felipe.contreras@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 12c4b90d542f4bb7b4c3f8228674d1ce1f3d55ee [formerly 12c4b90d542f4bb7b4c3f8228674d1ce1f3d55ee [formerly c01499ef6933bd0e3fc65277559dc80f13d6a49e]]
Former-commit-id: a2571123e8d524a0b66d472c3dfb0d33e83ee11b
Former-commit-id: 6fe29e271e992d4a0c1fc68dcbdd9cb1e437923a