gc: remove gc.pid file at end of execution
This file isn't really harmful, but isn't useful either, and can create
minor annoyance for the user:

* It's confusing, as the presence of a *.pid file often implies that a
  process is currently running. A user running "ls .git/" and finding
  this file may incorrectly guess that a "git gc" is currently running.

* Leaving this file means that a "git gc" in an already gc-ed repo is
  no-longer a no-op. A user running "git gc" in a set of repositories,
  and then synchronizing this set (e.g. rsync -av, unison, ...) will see
  all the gc.pid files as changed, which creates useless noise.

This patch unlinks the file after the garbage collection is done, so that
gc.pid is actually present only during execution.

Future versions of Git may want to use the information left in the gc.pid
file (e.g. for policies like "don't attempt to run a gc if one has
already been ran less than X hours ago"). If so, this patch can safely be
reverted. For now, let's not bother the users.

Explained-by: Matthieu Moy <Matthieu.Moy@imag.fr>
Signed-off-by: Jonathan Nieder <jrnieder@gmail.com>
Improved-by: Nguyễn Thái Ngọc Duy <pclouds@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 50d07947c6a5b934e7d03395f5ac989ce5935ddf [formerly 951a452685cd9f95c4c8eea594e45ee53a3dec4d] [formerly 951a452685cd9f95c4c8eea594e45ee53a3dec4d [formerly 4c5baf0273ae3382dbaca81fee0331cada5899ea]]
Former-commit-id: 45143edc3b88bb7fefc7e3af5595612cd24b2743
Former-commit-id: aa83dddcf82b64312f6b9262fc6db60365c20c0c