checkout: allow dwim for branch creation for "git checkout $branch --"
The "--" notation disambiguates files and branches, but as a side-effect
of the previous implementation, also disabled the branch auto-creation
when $branch does not exist.

A possible scenario is then:

git checkout $branch
=> fails if $branch is both a ref and a file, and suggests --

git checkout $branch --
=> refuses to create the $branch

This patch allows the second form to create $branch, and since the -- is
provided, it does not look for file named $branch.

Signed-off-by: Matthieu Moy <Matthieu.Moy@imag.fr>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 2a74e4246650967dc6f3e7654f898c273e6b4f1d [formerly 2a74e4246650967dc6f3e7654f898c273e6b4f1d [formerly a047fafc7866cc4087201e284dc1f53e8f9a32d5]]
Former-commit-id: 63d99cf018e42740dba373d329e77a2992f2fb0d
Former-commit-id: 853ea7f361a2bc10febf51a050cdd309f1ec9ede