Revert "test-lib: support running tests under valgrind in parallel"
This reverts commit 5fdd6db990a3531297a29f054af4881b9e116900 [formerly ad0e6233320b004f0d686f6887c803e508607bd2].

--valgrind-parallel was broken from the start: during review I made
the whole valgrind setup code conditional on not being a
--valgrind-parallel worker child.  But even the children crucially
need $GIT_VALGRIND to be set; it should therefore have been set
outside the conditional.

The fix would be a two-liner, but since the introduction of the
feature, almost four months have passed without anyone noticing that
it is broken.  So this feature is not worth the about hundred lines of
test-lib.sh complexity.  Revert it.

Signed-off-by: Thomas Rast <tr@thomasrast.ch>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 26a07309a6e195fc2d0adc89dcc06f3994ae7603