Documentation: revamp gitk(1)
The gitk manpage suffers from a bit of neglect: there have been only
minor changes, and no changes to the set of options documented, since
a2df1fb (Documentation: New GUI configuration and command-line
options., 2008-11-13).  In the meantime, the set of rev-list options
has been expanded several times by options that are useful in gitk,
e.g., --ancestry-path and the optional globbing for --branches, --tags
and --remotes.

Restructure and expand the manpage.  List more options that the author
perceives as useful, while remaining somewhat terse.  Ideally the user
should not have to look up any of the references, but we dispense with
precise explanations in some places and refer to git-log(1) instead.

Note that the options that have an easy GUI equivalent (e.g.,
--word-diff, -S, --grep) are deliberately not listed even in the cases
where they simply fill in the GUI fields.

Signed-off-by: Thomas Rast <trast@inf.ethz.ch>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 2bd2df44b77db5421235051e2ddbaf8d2c03298a [formerly 2bd2df44b77db5421235051e2ddbaf8d2c03298a [formerly 744db23c2d4399ff345a0ef6f2407c9a494d91dd]]
Former-commit-id: cd0f9690ee959ac630c0732683b2dab5f6e8df48
Former-commit-id: 2431162993d138683f15548e4c8e9bf35b3cb5ea