remote-hg: unquote C-style paths when exporting
git-fast-import documentation says that paths can be C-style quoted.
Unfortunately, the current remote-hg helper doesn't unquote quoted
path and pass them as-is to Mercurial when the commit is created.

This results in the following situation:

 - clone a mercurial repository with git
 - add a file with space in a directory: `>dir/foo\ bar`
 - commit that new file, and push the change to mercurial
 - the mercurial repository now has a new directory named '"dir',
   which contains a file named 'foo bar"'

Use Python str.decode('string-escape') to unquote the string if it
starts and ends with ".  It has been tested with quotes, spaces, and
utf-8 encoded file-names.

Signed-off-by: Antoine Pelisse <apelisse@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 63a71fe0a8c8258d529e8ef10ab05d2d8399d13c [formerly 63a71fe0a8c8258d529e8ef10ab05d2d8399d13c [formerly 113626537711e290af46548683e02ae0a2f51b6f]]
Former-commit-id: 18bef37483e53f7c553408ceef97359e05fdb2ef
Former-commit-id: 9ec900575281e7e6b0e5e5d22a9b35657c26978b