t5510: prepare test refs more straightforwardly
"git fetch" was being used with contrived refspecs to create tags and
remote-tracking branches in test repositories in preparation for the
actual tests.  This is obscure and also makes one wonder whether this
is indeed just preparation or whether some side-effect of "git fetch"
is being tested.

So use the more straightforward commands "git tag" / "git update-ref"
when preparing branches in test repositories.

Signed-off-by: Michael Haggerty <mhagger@alum.mit.edu>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 3ce7dbc1b7103acfa1e7311e98a5885e3ca7c6ee [formerly 3ce7dbc1b7103acfa1e7311e98a5885e3ca7c6ee [formerly d0d06e892a1a7bc6a1dfd76374da09f0170cae2c]]
Former-commit-id: 91d0c9ebf8297e7976c1d130e34acba33b4c2bd7
Former-commit-id: 677a3d6e9cac64cfd6a3a8ae9eb0ff89b708010b