t5510: check that "git fetch --prune --tags" does not prune branches
"git fetch --prune --tags" is currently interpreted as follows:

* "--tags" is equivalent to specifying a refspec
  "refs/tags/*:refs/tags/*", and supersedes any default refspecs
  configured via remote.$REMOTE.fetch.

* "--prune" only operates on the refspecs being fetched.

Therefore, "git fetch --prune --tags" prunes tags in refs/tags/* but
does not fetch or prune other references.  The fact that this command
does not prune references outside of refs/tags/* was previously
untested.  So add a test that verifies the status quo.

However, the status quo is surprising, so it will be changed later in
this patch series.

Signed-off-by: Michael Haggerty <mhagger@alum.mit.edu>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: c23962460b0e2d7e7f3e6c70093e13b7fe125618 [formerly 072b57930f06688505433deaf99faa1f908a53fd] [formerly 072b57930f06688505433deaf99faa1f908a53fd [formerly 68a304d5fed6657e452e8c1571fe9ca93d0a272a]]
Former-commit-id: 42d98a66164f7cc42e7e365f4ee19c5ff37768b7
Former-commit-id: 5043cd7959e68120cd2535139fa3f31b4199eec1