checkout_entry(): clarify the use of topath[] parameter
The said function has this signature:

	extern int checkout_entry(struct cache_entry *ce,
				  const struct checkout *state,
				  char *topath);

At first glance, it might appear that the caller of checkout_entry()
can specify to which path the contents are written out by the last
parameter, and it is tempting to add "const" in front of its type.

In reality, however, topath[] is to point at a buffer to store the
temporary path generated by the callchain originating from this
function, and the temporary path is always short, much shorter than
the buffer prepared by its only caller in builtin/checkout-index.c.

Document the code a bit to clarify so that future callers know how
to use the function better.

Noticed-by: Nguyễn Thái Ngọc Duy <pclouds@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 29f28312b71616034858216668d1de2ef1debb11 [formerly cb56bd3499ed879a6e6ec48dae52c51a7a5ea0cf] [formerly cb56bd3499ed879a6e6ec48dae52c51a7a5ea0cf [formerly af2a651d2ecb3967244c9a44a5813a0cf9977df5]]
Former-commit-id: bde44a31d90ac808df772d5f1fffd7bfe29fc5a7
Former-commit-id: 67f7804793b80aa701f2f63bbd8bdc644f8cb877