entry.c: convert checkout_entry to use strbuf
The old code does not do boundary check so any paths longer than
PATH_MAX can cause buffer overflow. Replace it with strbuf to handle
paths of arbitrary length.

The OS may reject if the path is too long though. But in that case we
report the cause (e.g. name too long) and usually move on to checking
out the next entry.

Signed-off-by: Nguyễn Thái Ngọc Duy <pclouds@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 359eb406ae9276bad166866d46da14e07bfdd474 [formerly 359eb406ae9276bad166866d46da14e07bfdd474 [formerly fd356f6aa8bb75ebef56fbc61caf7e02517fa6d3]]
Former-commit-id: 6a84862899dd22d35f485f7ae5b9086af346e746
Former-commit-id: 7322c3ff9c673c91bdc43b43fde44cd47461ce8f