log_tree_diff: die when we fail to parse a commit
We currently call parse_commit and then assume we can
dereference the resulting "tree" struct field. If parsing
failed, however, that field is NULL and we end up
segfaulting.

Instead of a segfault, let's print an error message and die
a little more gracefully.

Note that this should never happen in practice, but may
happen in a corrupt repository.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 963dbf6cc089877b297279f6bcab7c1512141e05 [formerly 2b9e8969b4a2be9d944a3ea7b1af2ec3227256cf] [formerly 2b9e8969b4a2be9d944a3ea7b1af2ec3227256cf [formerly 7059dccc6c60a872a314b19ac17702065a71d6bd]]
Former-commit-id: 32c5a381a0ccf5e9e0da40a7d60833d8e88461bd
Former-commit-id: 6d5268e9cf16d0d08925ebe142168109583284ca