checkout: do not die when leaving broken detached HEAD
If we move away from a detached HEAD that has broken or
corrupted commits, we might die in two places:

  1. Printing the "old HEAD was..." message.

  2. Printing the list of orphaned commits.

In both cases, we ignore the return value of parse_commit
and feed the resulting commit to the pretty-print machinery,
which will die() upon failing to read the commit object
itself.

Since both cases are ancillary to the real operation being
performed, let's be more robust and keep going. This lets
users more easily checkout away from broken history.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 8978be22dcb84c5cc231da8b7009d75bb9d98dfa [formerly 8978be22dcb84c5cc231da8b7009d75bb9d98dfa [formerly 3c62183929080c17299d5b404eb092e3d53c161a]]
Former-commit-id: 8120fb3ef1dccf36ebc0a3d51475b3a2eee2c661
Former-commit-id: ca3dbb138f5433960fd0212b7b4ce3b6089fce42