pack-objects: refactor the packing list
The hash table that stores the packing list for a given `pack-objects`
run was tightly coupled to the pack-objects code.

In this commit, we refactor the hash table and the underlying storage
array into a `packing_data` struct. The functionality for accessing and
adding entries to the packing list is hence accessible from other parts
of Git besides the `pack-objects` builtin.

This refactoring is a requirement for further patches in this series
that will require accessing the commit packing list from outside of
`pack-objects`.

The hash table implementation has been minimally altered: we now
use table sizes which are always a power of two, to ensure a uniform
index distribution in the array.

Signed-off-by: Vicent Marti <tanoku@gmail.com>
Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 5b5dc023cdeadd06ce0d0dc2e1549a5a766ffcb4 [formerly bf169d59e8712648acccbe8d9599a82a4fb72f60] [formerly bf169d59e8712648acccbe8d9599a82a4fb72f60 [formerly 2834bc27c16e4ae103768dd6dda530f0e46116af]]
Former-commit-id: 50edb103c6a660b0caf46a0a681edeb090ce3a6c
Former-commit-id: 399bc4aaa2901c0c62d1825afc3c01d51cd64ddf