revision: allow setting custom limiter function
This commit enables users of `struct rev_info` to peform custom limiting
during a revision walk (i.e. `get_revision`).

If the field `include_check` has been set to a callback, this callback
will be issued once for each commit before it is added to the "pending"
list of the revwalk. If the include check returns 0, the commit will be
marked as added but won't be pushed to the pending list, effectively
limiting the walk.

Signed-off-by: Vicent Marti <tanoku@gmail.com>
Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 00c89fe51117ba6240947db0f33867ab00ac89aa [formerly 00c89fe51117ba6240947db0f33867ab00ac89aa [formerly a330de31d18ca08c773c64e3657b3bafd59cf751]]
Former-commit-id: 860431466e34c45ee9c721f510e7c65eaa49705e
Former-commit-id: db3e3fe916785d0309b668f87697b4fa5ec3f0d2