http.c: Spell the null pointer as NULL
Commit 1bbcc224 ("http: refactor options to http_get_*", 28-09-2013)
changed the type of final 'options' argument of the http_get_file()
function from an int to an 'struct http_get_options' pointer.
However, it neglected to update the (single) call site. Since this
call was passing '0' to that argument, it was (correctly) being
interpreted as a null pointer. Change to argument to NULL.

Noticed by sparse. ("Using plain integer as NULL pointer")

Signed-off-by: Ramsay Jones <ramsay@ramsay1.demon.co.uk>
Acked-by: Jeff King <peff@peff.net>
Reviewed-by: Jonathan Nieder <jrnieder@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: fd61615e30dd40462666681e995503151b4f3f89 [formerly 600cd63f5232725ccead52b68117fcedf9dd1d53] [formerly 600cd63f5232725ccead52b68117fcedf9dd1d53 [formerly 70900eda4a0ed473ab5a933940285ac3dda698c0]]
Former-commit-id: e2b7a792b9e95eff4a4890432f86780be952bcb2
Former-commit-id: bdd6be3b7499d114c64c5b66a1cb528300710023