t3200: do not open a HTML manual page when DEFAULT_MAN_FORMAT is html
We have the build configuration option DEFAULT_MAN_FORMAT to choose a
format different from man pages to be used by 'git help' when no format
is requested explicitly. Since 65db0443 (Set the default help format to
html for msys builds, 2013-06-04) we use html on Windows by default.

There is one test in t3200-branch.sh that invokes a help page. The
intent of the redirections applied to the command invocation is to avoid
that the man page viewer interferes with the automated test. But when
the default format is not "man", this does not have the intended effect,
and the HTML manual page is opened during the test run. Request "man"
format explicitly to keep the test silent.

Signed-off-by: Johannes Sixt <j6t@kdbg.org>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 7007736de6dfaa3a48320adaf2ae6c531bb58d28 [formerly a3ceac18520e3c14721be4901f1113aea5d5cd5a] [formerly a3ceac18520e3c14721be4901f1113aea5d5cd5a [formerly 01e8d327a9d647c1e891920365c5f6b4f595d5e9]]
Former-commit-id: ac3d268c9ed8f027c97702bfaa4c358dc91e9341
Former-commit-id: f471be26ba10342d5139dd1fef0bb044b8fc1b2d