Windows: a test_cmp that is agnostic to random LF <> CRLF conversions
In a number of tests, output that was produced by a shell script is
compared to expected output using test_cmp. Unfortunately, the MSYS bash--
when invoked via git, such as in hooks--converts LF to CRLF on output
(as produced by echo and printf), which leads to many false positives.

Implements a diff tool that undoes the converted CRLF. To avoid that
sub-processes are spawned (which is very slow on Windows), the tool is
implemented as a shell function. Diff is invoked as usual only when a
difference is detected by the shell code.

Signed-off-by: Johannes Sixt <j6t@kdbg.org>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 8f54dc1ab39e4087fd2c2fc3ccb7a1aa7e33a239 [formerly 8f54dc1ab39e4087fd2c2fc3ccb7a1aa7e33a239 [formerly 4d715ac05cf5b800a98ce75bbc0cb3ea2dfe70ab]]
Former-commit-id: 46f2e7991ad3a89b8eb69c758e7d5b4b981b6d48
Former-commit-id: 25acb513807006ec8282aeeb5a6c58df90afc081