sha1_file: move comment about return value where it belongs
Commit 5b0864070 (sha1_object_info_extended: make type calculation
optional, Jul 12 2013) changed the return value of the
sha1_object_info_extended function to 0/-1 for success/error.

Previously this function returned the object type for success or
-1 for error. But unfortunately the above commit forgot to change
or move the comment above this function that says "returns enum
object_type or negative".

To fix this inconsistency, let's move the comment above the
sha1_object_info function where it is still true.

Signed-off-by: Christian Couder <chriscool@tuxfamily.org>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 5c44802cea4efb12003e0f8a6a2bfdd7e45b5c55 [formerly 5c44802cea4efb12003e0f8a6a2bfdd7e45b5c55 [formerly 3fc0dca9ceb653798377591018c81324f4d84032]]
Former-commit-id: 8a0e3d3655ab1d4b946d365d16bb7f9b30d5167c
Former-commit-id: 98595e7ead22dd312263b8830c47dfc84b2a1517