t/README: tests can use perl even with NO_PERL
The git build system supports a NO_PERL switch to avoid installing
perl bindings or other features (like "git add --patch") that rely on
perl on runtime, but even with NO_PERL it has not been possible for a
long time to run tests without perl.  Helpers such as

	nul_to_q () {
		"$PERL_PATH" -pe 'y/\000/Q/'
	}

use perl as a better tr or sed and are regularly used in tests without
worrying to add a PERL prerequisite.

Perl is portable enough that it seems fine to keep relying on it for
this kind of thing in tests (and more readable than the alternative of
trying to find POSIXy equivalents).  Update the test documentation to
clarify this.

Reported-by: Johannes Sixt <j6t@kdbg.org>
Signed-off-by: Jonathan Nieder <jrnieder@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: f2ce5a81fdc2bdde3d15b7a9fcd9a2fe8ca3a5ac [formerly 8744cf257862ebd4d282252b75ad537ba1e2686b] [formerly 8744cf257862ebd4d282252b75ad537ba1e2686b [formerly f8fc0ee314d0e0c993d2edac8848a4cfe72c771f]]
Former-commit-id: fcc9dafd4007fc61fc78e336f768caf32199feee
Former-commit-id: e87aba2fb016e29a49974df33213d7f83240a231