use @@PERL@@ in built scripts
Several of the built shell commands invoke a bare "perl" to
perform some one-liners. This will use the first perl in the
PATH rather than the one specified by the user's SHELL_PATH.
We are not asking these perl invocations to do anything
exotic, so typically any old system perl will do; however,
in some cases the system perl may have unexpected behavior
(e.g., by handling line endings differently). We should err
on the side of using the perl the user pointed us to.

The downside of this is that on systems with a sane perl
setup, we no longer find the perl at runtime, but instead
point to a static perl (like /usr/bin/perl). That means we
will not handle somebody moving perl without rebuilding git,
whereas before we tracked it just fine. This is probably not
a big deal, though, as the built perl scripts already
suffered from this.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 7745daa1b77f92cf406fa6674c55bcc9d3e21262 [formerly 7745daa1b77f92cf406fa6674c55bcc9d3e21262 [formerly fcb06a8d54e29d12bb8160b978ce78331c95a1bf]]
Former-commit-id: a1789cd8ae731ffe174ee3f136a5630b8160050d
Former-commit-id: bdc069a49225d4344c901fced11bc68261281770