t: use perl instead of "$PERL_PATH" where applicable
As of the last commit, we can use "perl" instead of
"$PERL_PATH" when running tests, as the former is now a
function which uses the latter. As the shorter "perl" is
easier on the eyes, let's switch to using it everywhere.

This is not quite a mechanical s/$PERL_PATH/perl/
replacement, though. There are some places where we invoke
perl from a script we generate on the fly, and those scripts
do not have access to our internal shell functions. The
result can be double-checked by running:

  ln -s /bin/false bin-wrappers/perl
  make test

which continues to pass even after this patch.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: adab6b69cae30a4b2f72c4b964f0bd6022793d45 [formerly adab6b69cae30a4b2f72c4b964f0bd6022793d45 [formerly 94221d22036f36e10680c0a1e7eafb4bceeb1248]]
Former-commit-id: 666d12dc7841eba33a6b38ae07fda1d81c9e0774
Former-commit-id: da78aaa1964f2da6cc730ce328fb264dc07090fa