subtree: add makefile target for html docs
The Makefile currently builds the roff manpage, but not the
html form. As some people may prefer the latter, let's make
it an option to build that, too. We also wire it into "make
doc" so that it is built by default.

This patch does not build or install it as part of
"install-doc"; that would require extra infrastructure to
handle installing the html as we do in git's regular
Documentation/ tree. That can come later if somebody is
interested.

Tested-by: Sebastian Schuberth <sschuberth@gmail.com>
Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 86bb1583dcf3d8f938e10d9bf6fc63f40ce6a901 [formerly 86bb1583dcf3d8f938e10d9bf6fc63f40ce6a901 [formerly 95c62fb9ea35fab9706792e858ce66967501eb00]]
Former-commit-id: ecab6af383b22d6779c03229768856b60e1c0500
Former-commit-id: d2a66e8140029a482cbfbacaa9b96dabeca27fd0