fetch --tags: fetch tags *in addition to* other stuff
Previously, fetch's "--tags" option was considered equivalent to
specifying the refspec "refs/tags/*:refs/tags/*" on the command line;
in particular, it caused the remote.<name>.refspec configuration to be
ignored.

But it is not very useful to fetch tags without also fetching other
references, whereas it *is* quite useful to be able to fetch tags *in
addition to* other references.  So change the semantics of this option
to do the latter.

If a user wants to fetch *only* tags, then it is still possible to
specifying an explicit refspec:

    git fetch <remote> 'refs/tags/*:refs/tags/*'

Please note that the documentation prior to 1.8.0.3 was ambiguous
about this aspect of "fetch --tags" behavior.  Commit

    cf9035f1b5 [formerly 1518298f98] [formerly 5ae30cf52b] [formerly af6021f30f [formerly af6021f30f [formerly f0cb2f137c]]] 2012-12-14 fetch --tags: clarify documentation

made the documentation match the old behavior.  This commit changes
the documentation to match the new behavior.

Signed-off-by: Michael Haggerty <mhagger@alum.mit.edu>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 3087aa5ddd8139ddc3d9baeb5ae43f9ee0c57d32 [formerly 3087aa5ddd8139ddc3d9baeb5ae43f9ee0c57d32 [formerly c5a84e92a2fe9e8748e32341c344d7a6c0f52a50]]
Former-commit-id: eeac7f5dcdf5ec292f10b586cffaaf82ddfaad5f
Former-commit-id: b089e32bf0078a567dcf272180eb01908f26287f