fetch, remote: properly convey --no-prune options to subprocesses
If --no-prune is passed to one of the following commands:

    git fetch --all
    git fetch --multiple
    git fetch --recurse-submodules
    git remote update

then it must also be passed to the "fetch" subprocesses that those
commands use to do their work.  Otherwise there might be a fetch.prune
or remote.<name>.prune configuration setting that causes pruning to
occur, contrary to the user's express wish.

Signed-off-by: Michael Haggerty <mhagger@alum.mit.edu>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 6fb06ffad3e0e5a21e584b13e686a127b2e047c3 [formerly 6fb06ffad3e0e5a21e584b13e686a127b2e047c3 [formerly 90765fa3e01658ac2fb75a4eb740cee7dad4a2dc]]
Former-commit-id: 78abc6e5419d4f385230d9a0afce3b3237a1bb97
Former-commit-id: 2384215cd9b556b3b2acf209af8387731fde4035