git-fetch.txt: improve description of tag auto-following
Make it clearer that tags are fetched independent of which branches
were fetched from the remote in any particular fetch.  (Tags are even
fetched if they point at objects that are in the current repository
but not reachable, which is probably a bug.)

Put less emphasis on the mechanism and more on the effect of tag
auto-following.  Also mention the options and configuration settings
that can change the tag-fetching behavior.

Signed-off-by: Michael Haggerty <mhagger@alum.mit.edu>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 265a327ad6d008952816a4888497adea2e46e8ca [formerly 265a327ad6d008952816a4888497adea2e46e8ca [formerly 37f0dcbdc1585dafd81c393aa750292e8634035d]]
Former-commit-id: 9786d18fd101181d12224edac5d4376738b85187
Former-commit-id: 99beec2d8dc091f30e28eb06f85520c19d437c58