fetch: improve the error messages emitted for conflicting refspecs
If we find two refspecs that want to update the same local reference,
emit an error message that is more informative based on whether one of
the conflicting refspecs is an opportunistic update during a fetch
with explicit command-line refspecs.  And especially, do not die if an
opportunistic reference update conflicts with an express wish of the
user; rather, just emit a warning and skip the opportunistic reference
update.

Signed-off-by: Michael Haggerty <mhagger@alum.mit.edu>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 81cea015d543f845b284a3fb3f88721355f70569 [formerly 8e2921d932825361afd43aa3c9a6bce8f686a0bc] [formerly 8e2921d932825361afd43aa3c9a6bce8f686a0bc [formerly f096e6e826678c29e4bfde4d9d1ae1df79074ce3]]
Former-commit-id: 028abb2ee1cc1039d93dfc555c3920db89247329
Former-commit-id: 0d72ae9a7cfa311889613680b1b177fb5b9cdb2c