http: return curl's AUTHAVAIL via slot_results
Callers of the http code may want to know which auth types
were available for the previous request. But after finishing
with the curl slot, they are not supposed to look at the
curl handle again. We already handle returning other
information via the slot_results struct; let's add a flag to
check the available auth.

Note that older versions of curl did not support this, so we
simply return 0 (something like "-1" would be worse, as the
value is a bitflag and we might accidentally set a flag).
This is sufficient for the callers planned in this series,
who only trigger some optional behavior if particular bits
are set, and can live with a fake "no bits" answer.

Signed-off-by: Jeff King <peff@peff.net>

Former-commit-id: 2cb37b3e72b1d43903c37fd129c66efcd4ac3bc3 [formerly 2cb37b3e72b1d43903c37fd129c66efcd4ac3bc3 [formerly 0972ccd97cc8f913ea828a1e03ef3652fc1ff514]]
Former-commit-id: 01778accc26cc7e65d8e62cf827098dfa074a0a1
Former-commit-id: 4cfc88b7005cd5e281fc2366b0641af20f6374d5