Fix '\%o' for printf from coreutils
The printf utility provided by coreutils when interpreting '\%o' format
does not recognize %o as formatting directive. For example
printf '\%o 0 returns \%o and warning: ignoring excess arguments,
starting with ‘0’, which results in failed tests in
t5309-pack-delta-cycles.sh. In most shells the test ends with success as
the printf is a builtin utility.

Fix it by using '\\%o' which is interpreted consistently in all versions
of printf.

Signed-off-by: Kacper Kornet <draenog@pld-linux.org>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 4d9e29642b931bf9305e787ae3fcc99fa3255e74 [formerly 4d9e29642b931bf9305e787ae3fcc99fa3255e74 [formerly fbc812a6009ebc91f4870c3572821cb4d68184f8]]
Former-commit-id: 3e05629b4541e8a56a8e9c53d2a257900ceebc45
Former-commit-id: 2ed9adf3066e06a09d7972c6f188f9879bc042c5