doc/pull: clarify the illustrations
The second illustration that shows the history after "git pull"
spelled the remote-tracking branch with "remotes/" prefix, which
is not necessary.  Drop it.

To match the assumption that a remote-tracking branch is used to
keep track of the advancement of the master at the origin, update
the first illustration that shows the history before "git pull"
to show the distinction between the master currently at origin and
the stale origin/master remote-tracking branch.

Noticed-by: Felipe Contreras <felipe.contreras@gmail.com>
Helped-by: Max Horn <max@quendi.de>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 2c12824af765998472fd99b29c9b2611d37dc7a9 [formerly 2c12824af765998472fd99b29c9b2611d37dc7a9 [formerly 5a3fd6afd4a26accc20a58bf4980f296772be98a]]
Former-commit-id: 0e754361c417192f0ff5981221dbd25a042f4693
Former-commit-id: f592110636e23a29630cb2a30277395d7d883a29