rev-list --exclude: tests
Add tests for the --exclude=<glob> feature.

A few tests are added for cases where use of globbing and
"--exclude" results in no positive revisions:

 * "--exclude=<glob>" before "--all" etc. resulted in no results;

 * "--stdin" is used but no input was given;

 * "--all" etc. is used but no matching refs are found.

Currently, we fail such a request with the same error message we
would give to a command line that does not specify any positive
revision (e.g. "git rev-list<ENTER>").

We may want to treat these cases differently and not error out, but
the logic to detect that would be common to all of them, so I'd
leave it outside this topic for now, and stop at adding these tests
as food-for-thought.

Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 7972a08ec92755cb3acae87d6aa167fff254b1cf [formerly 254dfcd6a57094ced7167d36b39393336b9f7e4b] [formerly 254dfcd6a57094ced7167d36b39393336b9f7e4b [formerly 751a2ac6edfbc803b9ed7c48d2d1fc54c97cc64c]]
Former-commit-id: 5f19e6b25ef79813e3fddaefe4b4a58eb17863ba
Former-commit-id: a0ce66e3f347a039ea10bcd8eefda4b477d813bc