submodule: don't access the .gitmodules cache entry after removing it
Commit 1c80a2a254b [formerly 07e1f3e2a33] [formerly b869453d501] [formerly 72ced407665 [formerly 72ced407665 [formerly 5fee995244e]]] introduced the stage_updated_gitmodules() function to
add submodule configuration updates to the index. It assumed that even
after calling remove_cache_entry_at() the same cache entry would still be
valid. This was true in the old days, as cache entries could never be
freed, but that is not so sure in the present as there is ongoing work to
free removed cache entries, which makes this code segfault.

Fix that by calling add_file_to_cache() instead of open coding it. Also
remove the "could not find .gitmodules in index" warning, as that won't
happen in regular use cases (and by then just silently adding it to the
index we do the right thing).

Thanks-to: Karsten Blees <karsten.blees@gmail.com>
Signed-off-by: Jens Lehmann <Jens.Lehmann@web.de>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 56d9f7f8b93d717606c94a969ceae1c017a934ca [formerly 56d9f7f8b93d717606c94a969ceae1c017a934ca [formerly bc8d6b9b904350c72017a3c92fdacbfba5b1be3f]]
Former-commit-id: ca8d008b4fa8f7969eaed4a1862e9f65dbccd711
Former-commit-id: e89491ae9e47689398a55445d39a58b2074b442b