submodule update: remove unnecessary orig_flags variable
cmd_update() in the submodule script tries to preserve the options given
on the command line in the "orig_flags" variable to pass them on into the
recursion when the '--recursive' option is given. But this isn't necessary
because all the variables set by the options will be seen in the recursion
too as that is achieved by executing "eval cmd_update".

The same has already been done for cmd_status() in e15bec0ec, so let's
clean up cmd_update() likewise. Also add a test to make sure that a
submodule name given on the command line is not passed into the recursion
(which was the goal of adding the orig_flags variable in 98dbe63db).

Signed-off-by: Jens Lehmann <Jens.Lehmann@web.de>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 4a774aa32cf4d38e5b3dc546d3bd982b5319ee09 [formerly 4a774aa32cf4d38e5b3dc546d3bd982b5319ee09 [formerly 361412828ab580e145c3350cf075b283ebe331a5]]
Former-commit-id: 02703cd5a7899b5cb8a8d9ee9a8dbd33db57e855
Former-commit-id: 0319bd4a0d605bedfc8857c16ad52bc20b621642