config: arbitrary number of matches for --unset and --replace-all
git-config used a static match array to hold the matches we want to
unset/replace when using --unset or --replace-all.  Use a
variable-sized array instead.

This in particular fixes the symptoms git-svn had when storing large
numbers of svn-remote.*.added-placeholder entries in the config file.

While the tests are rather more paranoid than just --unset and
--replace-all, the other operations already worked.  Indeed git-svn's
usage only breaks the first time *after* creating so many entries,
when it wants to unset and re-add them all.

Reported-by: Jess Hottenstein <jess.hottenstein@gmail.com>
Signed-off-by: Thomas Rast <tr@thomasrast.ch>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 75ab571d541f709d4027aa5cc8577a81b7840f95 [formerly 75ab571d541f709d4027aa5cc8577a81b7840f95 [formerly 83786fa412662f6d7cdade9e4043882c5ad21c1a]]
Former-commit-id: 6dd5fdf50eea233278bd7e59d44f14d2e6696fa8
Former-commit-id: 31b4b6485b636df061c307c766e97dd1e5378de9