compat: add endianness helpers
The POSIX standard doesn't currently define a `ntohll`/`htonll`
function pair to perform network-to-host and host-to-network
swaps of 64-bit data. These 64-bit swaps are necessary for the on-disk
storage of EWAH bitmaps if they are not in native byte order.

Many thanks to Ramsay Jones <ramsay@ramsay1.demon.co.uk> and
Torsten Bögershausen <tboegi@web.de> for cygwin/mingw/msvc
portability fixes.

Signed-off-by: Vicent Marti <tanoku@gmail.com>
Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: f05be02bdebda3784dfe236bd54864045c962b34 [formerly 6ba40eceabdaeb0601e41dea3ed1e3811753a112] [formerly 6ba40eceabdaeb0601e41dea3ed1e3811753a112 [formerly 7e3dae494370b5596a6ea76af1191829ce11bce2]]
Former-commit-id: 00f92dc8701d73986d563c1106ce1b22610afb21
Former-commit-id: b7d35c263c5b6b60b9c18c6816a0986711ab2a40