branch: fix --verbose output column alignment
Commit f2e0873 (branch: report invalid tracking branch as gone) removed
an early return from fill_tracking_info() in the path taken when 'git
branch -v' lists a branch in sync with its upstream. This resulted in an
unconditionally added space in front of the subject line:

    $ git branch -v
    * master f5eb3da  commit pushed to upstream
      topic  f935eb6 unpublished topic

Instead, only add the trailing space if a decoration have been added.

To catch this kind of whitespace breakage in the tests, be a bit less
smart when filtering the output through sed.

Signed-off-by: Torstein Hegge <hegge@resisty.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: adfe4d190a7a2158d4a9050c975ee6fafe3aa901 [formerly adfe4d190a7a2158d4a9050c975ee6fafe3aa901 [formerly 6b364d48f2e72d6c37115e6aa4fe769a523dd1d4]]
Former-commit-id: b0a15e14651bcb8835952b6a7dcbbb45970ab23b
Former-commit-id: 36f811ad56a7cd5f001c9ba528387339833ce902