gitk: Split out diff part in $commitinfo
So far we just parsed everything after the headers into the "comment"
bit of $commitinfo, including notes and -- if you gave weird options
-- the diff.

Split out the diff, if any, into a separate field.  It's easy to
recognize, since it always starts with /^diff/ and is preceded by an
empty line.

We take care to snip away said empty line.  The display code already
properly spaces the end of the message from the first diff, and
leaving another empty line at the end looks ugly.

Signed-off-by: Thomas Rast <trast@inf.ethz.ch>
Signed-off-by: Paul Mackerras <paulus@samba.org>

Former-commit-id: 5014f7d007244b859790b05fdd7cf907a64f2b2b [formerly 5014f7d007244b859790b05fdd7cf907a64f2b2b [formerly b449eb2cb34fbdcb5cab7101a527985013818c9b]]
Former-commit-id: 486299b385d75af98a83acad75368854ac0bd12b
Former-commit-id: 81003f800260533adf55c1f62cbf9c9fb3b06760