gitk: Recognize -L option
This gives line-log support to gitk, by exploiting the new support for
processing and showing "inline" diffs straight from the git-log
output.

Note that we 'set allknown 0', which is a bit counterintuitive since
this is a "known" option.  But that flag prevents gitk from thinking
it can optimize the view by running rev-list to see the topology; in
the -L case that doesn't work.

Signed-off-by: Thomas Rast <trast@inf.ethz.ch>
Signed-off-by: Paul Mackerras <paulus@samba.org>

Former-commit-id: 7e711185e405044d44cd5c09f0355e0e3ca34bd3 [formerly 7e711185e405044d44cd5c09f0355e0e3ca34bd3 [formerly ce2c58cdaad42b9af6ad243f149c542ea93f8d01]]
Former-commit-id: 5e15f41a23e5f85bb82366f61e9f0ae1a5d03e26
Former-commit-id: db992dd3b066038dbd53ebdbccc4c4be4526252f