test-hg.sh: eliminate 'local' bashism
Unlike bash, POSIX shell does not specify a 'local' command for
declaring function-local variable scope.  Except for IFS, the variable
names are not used anywhere else in the script so simply remove the
'local'.  For IFS, move the assignment to the 'read' command to
prevent it from affecting code outside the function.

Signed-off-by: Richard Hansen <rhansen@bbn.com>
Reviewed-by: Felipe Contreras <felipe.contreras@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: d6b989ac1f3d8a41b0d465a2b841be28ec5b91f0 [formerly d6b989ac1f3d8a41b0d465a2b841be28ec5b91f0 [formerly 5105edd411bc472015d63eadaf1f826bebdc92e5]]
Former-commit-id: debad67080a6fc582367afff7a9547cd76d0d507
Former-commit-id: f3b89069c827fda9d29bddc696bd67cc6fb6614a