test: replace shebangs with descriptions in shell libraries
A #! line in these files is misleading, since these scriptlets are
meant to be sourced with '.' (using whatever shell sources them)
instead of run directly using the interpreter named on the #! line.

Removing the #! line shouldn't hurt syntax highlighting since
these files have filenames ending with '.sh'.  For documentation,
add a brief description of how the files are meant to be used in
place of the shebang line.

Signed-off-by: Jonathan Nieder <jrnieder@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 1e1d02d9dbb25e5b0cc3f3be1074cb0a58f2ba12 [formerly 1e1d02d9dbb25e5b0cc3f3be1074cb0a58f2ba12 [formerly c74c72034f061d1d9d5b8b1fba20ce0138d423b4]]
Former-commit-id: b84665fa27c488f95a1265e668e359bdb0427040
Former-commit-id: 8e3343f662296c9730dbe2251fae6cc98303bc78