git_connect: remove artificial limit of a remote command
Since day one, function git_connect() had a limit on the command line of
the command that is invoked to make a connection. 7a33bcbe converted the
code that constructs the command to strbuf. This would have been the
right time to remove the limit, but it did not happen. Remove it now.

git_connect() uses start_command() to invoke the command; consequently,
the limits of the system still apply, but are diagnosed only at execve()
time. But these limits are more lenient than the 1K that git_connect()
imposed.

Signed-off-by: Johannes Sixt <j6t@kdbg.org>
Signed-off-by: Torsten Bögershausen <tboegi@web.de>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 353799c9d3dc9057440c7357f506ab4689d5aae4 [formerly 353799c9d3dc9057440c7357f506ab4689d5aae4 [formerly d98d109979760bed1909fbf5e56cbb6239980095]]
Former-commit-id: 94535977e2176e7359126093761ab449ed3f4baf
Former-commit-id: 8f8afacaf3b82e0b08e59b2962bfd0875823e7ac