replace {pre,suf}fixcmp() with {starts,ends}_with()
Leaving only the function definitions and declarations so that any
new topic in flight can still make use of the old functions, replace
existing uses of the prefixcmp() and suffixcmp() with new API
functions.

The change can be recreated by mechanically applying this:

    $ git grep -l -e prefixcmp -e suffixcmp -- \*.c |
      grep -v strbuf\\.c |
      xargs perl -pi -e '
        s|!prefixcmp\(|starts_with\(|g;
        s|prefixcmp\(|!starts_with\(|g;
        s|!suffixcmp\(|ends_with\(|g;
        s|suffixcmp\(|!ends_with\(|g;
      '

on the result of preparatory changes in this series.

Signed-off-by: Christian Couder <chriscool@tuxfamily.org>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: b662a41f4b75161f615860aadad7b7d2c1769d78 [formerly b662a41f4b75161f615860aadad7b7d2c1769d78 [formerly 59556548230e617b837343c2c07e357e688e2ca4]]
Former-commit-id: 8575bcb18139ac4f2240fd26780ed3e657fb9ddf
Former-commit-id: 380686979f62c804e2ef5904e0d08eabacdc3147