gettext.c: detect the vsnprintf bug at runtime
Bug 6530 [1] in glibc causes "git show v0.99.6~1" to fail with error
"your vsnprintf is broken". The workaround avoids that, but it
corrupts system error messages in non-C locales.

The bug has been fixed since 2.17. We could know running glibc version
with gnu_get_libc_version(). But version is not a sure way to detect
the bug because downstream may back port the fix to older versions. Do
a runtime test that immitates the call flow that leads to "your
vsnprintf is broken". Only enable the workaround if the test fails.

Tested on Gentoo Linux, glibc 2.16.0 and 2.17, amd64.

[1] http://sourceware.org/bugzilla/show_bug.cgi?id=6530

Signed-off-by: Nguyễn Thái Ngọc Duy <pclouds@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 98dc8c2da7bab223ca98baeba0d683a21c6b5208 [formerly 98dc8c2da7bab223ca98baeba0d683a21c6b5208 [formerly 9c0495d23e6999375976ca44e3812fc65b73626e]]
Former-commit-id: 18accaa893fb55f08202fc11bf45b37418f5d396
Former-commit-id: ba40f3a02f0ffadaef6473e90d2593347c27671c