builtin/remote: remove postfixcmp() and use suffixcmp() instead
Commit 8cc5b290 (git merge -X<option>, 25 Nov 2009) introduced
suffixcmp() with nearly the same implementation as postfixcmp()
that already existed since commit 211c8968 (Make git-remote a
builtin, 29 Feb 2008).

The only difference between the two implementations is that,
when the string is smaller than the suffix, one implementation
returns 1 while the other one returns -1.

But, as postfixcmp() is only used to compare for equality, the
distinction does not matter and does not affect the correctness of
this patch.

As postfixcmp() has always been static in builtin/remote.c
and is used nowhere else, it makes more sense to remove it
and use suffixcmp() instead in builtin/remote.c, rather than
to remove suffixcmp().

Reviewed-by: Jonathan Nieder <jrnieder@gmail.com>
Signed-off-by: Christian Couder <chriscool@tuxfamily.org>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: b3e8cece44f1f846ea8ca7e4725c805f99563e54 [formerly 57fc71b8780d04699ed0a23de8a3c506889051dd] [formerly 57fc71b8780d04699ed0a23de8a3c506889051dd [formerly 3fb5aead2971c6181dbf143295c427175b8bf082]]
Former-commit-id: 8273a340d240a076fdd857ad793840d90c0eb81e
Former-commit-id: 22ca1bb5d811714c1018772a7cca1fce0ce5a359