strbuf: introduce starts_with() and ends_with()
prefixcmp() and suffixcmp() share the common "cmp" suffix that
typically are used to name functions that can be used for ordering,
but they can't, because they are not antisymmetric:

        prefixcmp("foo", "foobar") < 0
        prefixcmp("foobar", "foo") == 0

We in fact do not use these functions for ordering.  Replace them
with functions that just check for equality.

Add starts_with() and end_with() that will be used to replace
prefixcmp() and suffixcmp(), respectively, as the first step.  These
are named after corresponding functions/methods in programming
languages, like Java, Python and Ruby.

In vcs-svn/fast_export.c, there was already an ends_with() function
that did the same thing. Let's use the new one instead while at it.

Signed-off-by: Christian Couder <chriscool@tuxfamily.org>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 517a174f21aa6ab40af72229e7cc594e39d68b95 [formerly 83fa93eb28daefc63a6fa03a1ad38a7b864203af] [formerly 83fa93eb28daefc63a6fa03a1ad38a7b864203af [formerly 956623157f828b2b4fd91a9bc5e78ba8e42437d9]]
Former-commit-id: 236568bef2de1ae8b63281ca57badc7eaa4399da
Former-commit-id: 9e298473e99ad7d42767dfccfd5d35b1164588ef