submodule: do not copy unknown update mode from .gitmodules
When submodule.$name.update is given as hint from the upstream in
the .gitmodules file, we used to blindly copy it to .git/config,
unless there already is a value defined for the submodule.

However, there is no reason to expect that the update mode hinted by
the upstream is available in the version of Git the user is using,
and a really custom "!cmd" prepared by an upstream person running on
Linux may not even be available to a user on Windows.  It is simply
irresponsible to copy the setting blindly and to attempt to use it
during a later "submodule update" without validating it first.

Just show the suggested value to the diagnostic output, and set the
value to 'none' in the configuration, if it is not one of the ones
that are known to be supported by this version of Git.

Helped-by: Jens Lehmann <Jens.Lehmann@web.de>
Helped-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 10ca004dbf5f3b410f45f9504932a5abdfccf26b [formerly 10ca004dbf5f3b410f45f9504932a5abdfccf26b [formerly ac1fbbda2013416b6c6a93d65c5dcf6662a60579]]
Former-commit-id: a2b18222b047c65912a5843b98b1559cd82d1c0e
Former-commit-id: 194a28aa91b806725b5d875e6d49557fb3389f66