mv: let 'git mv file no-such-dir/' error out
Git used to trim the trailing slash, and make the command equivalent
to 'git mv file no-such-dir', which created the file no-such-dir
(while the trailing slash explicitly stated that it could only be a
directory).

This patch skips the trailing slash removal for the destination
path.  The path with its trailing slash is passed to rename(2),
which errors out with the appropriate message:

  $ git mv file no-such-dir/
  fatal: renaming 'file' failed: Not a directory

Original-patch-by: Duy Nguyen <pclouds@gmail.com>
Signed-off-by: Matthieu Moy <Matthieu.Moy@imag.fr>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: d05d6b1abad821c281c15efc6f0790ace886c37a [formerly d05d6b1abad821c281c15efc6f0790ace886c37a [formerly c57f6281ff847db7094d708f5c5b939d378f7120]]
Former-commit-id: 864464300fea1b4674bef0f101b36602ef73d53f
Former-commit-id: 937807203f8b16d6e93cb5efa3b1b1b4ee69a5d8