push: also use "upstream" mapping when pushing a single ref
When the user is using the 'upstream' mode, these commands:

    $ git push
    $ git push origin

would find the 'upstream' branch for the current branch, and then
push the current branch to update it.  However, pushing a single
branch explicitly, i.e.

    $ git push origin $(git symbolic-ref --short HEAD)

would not go through the same ref mapping process, and ends up
updating the branch at 'origin' of the same name, which may not
necessarily be the upstream of the branch being pushed.

In the spirit similar to the previous one, map a colon-less refspec
using the upstream mapping logic.

Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 5d552da8746a367c7d00d4f7de1f27ba6142e280 [formerly 5d552da8746a367c7d00d4f7de1f27ba6142e280 [formerly fc9261ca611080b1dae76b86b3bf5f36d592042f]]
Former-commit-id: a7f56155551e35659570fb924bd76e498bba7115
Former-commit-id: b750ee062754b0da4977a75a6d8807e0e421f5ad