clone: prevent --reference to a shallow repository
If we borrow objects from another repository, we should also pay
attention to their $GIT_DIR/shallow (and even info/grafts). But
current alternates code does not.

Reject alternate repos that are shallow because we do not do it
right. In future the alternate code may be updated to check
$GIT_DIR/shallow properly so that this restriction could be lifted.

Signed-off-by: Nguyễn Thái Ngọc Duy <pclouds@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: ae4f357f5c24493af60b00a355d87319e2f757d2 [formerly ae4f357f5c24493af60b00a355d87319e2f757d2 [formerly 606e435a0a11634744282068174240caf2926fec]]
Former-commit-id: 557294409718034f5899fff79add62772b923705
Former-commit-id: 562acc14e9d8a57249706eec9ed472b9c31b63fb