upload-pack: make sure deepening preserves shallow roots
When "fetch --depth=N" where N exceeds the longest chain of history in
the source repo, usually we just send an "unshallow" line to the
client so full history is obtained.

When the source repo is shallow we need to make sure to "unshallow"
the current shallow point _and_ "shallow" again when the commit
reaches its shallow bottom in the source repo.

This should fix both cases: large <N> and --unshallow.

Signed-off-by: Nguyễn Thái Ngọc Duy <pclouds@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: ab7919fa5c6ec7a498f85c32f9777706c1d986f0 [formerly af288b3a1891bc8e8a03bc68fb2fb739b6a9444e] [formerly af288b3a1891bc8e8a03bc68fb2fb739b6a9444e [formerly 79d3a236c551ad59719a6835bee03a1446296e65]]
Former-commit-id: 732dd47bdbbf826917a582086d16a47579964c5d
Former-commit-id: bf53b3f95679f5d54c920747b72dcccfb4bd4c3e