add GIT_SHALLOW_FILE to propagate --shallow-file to subprocesses
This may be needed when a hook is run after a new shallow pack is
received, but .git/shallow is not settled yet. A temporary shallow
file to plug all loose ends should be used instead. GIT_SHALLOW_FILE
is overriden by --shallow-file.

--shallow-file does not work in this case because the hook may spawn
many git subprocesses and the launch commands do not have
--shallow-file as it's a recent addition.

Signed-off-by: Nguyễn Thái Ngọc Duy <pclouds@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 68be4279fb22dcd10ce0a0783b186e79e2d81d02 [formerly 38755b0d66a8a1a342aa4c840c8f563c2a0533f5] [formerly 38755b0d66a8a1a342aa4c840c8f563c2a0533f5 [formerly 069c053222bfc62a6522430a137e9b2c7ff36e4c]]
Former-commit-id: 8b6c9981a0083512a98275430583abcd3e5b2d0a
Former-commit-id: 17253655c1e7fd17456ef767247848f2f1a0d88a