clone: use git protocol for cloning shallow repo locally
clone_local() does not handle $SRC/shallow. It could be made so, but
it's simpler to use fetch-pack/upload-pack instead.

This used to be caught by the check in upload-pack, which is triggered
by transport_get_remote_refs(), even in local clone case. The check is
now gone and check_everything_connected() should catch the result
incomplete repo. But check_everything_connected() will soon be skipped
in local clone case, opening a door to corrupt repo. This patch should
close that door.

Signed-off-by: Nguyễn Thái Ngọc Duy <pclouds@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 626626a902836a1b3420c7e35eb69c6d00026277 [formerly 626626a902836a1b3420c7e35eb69c6d00026277 [formerly 0d7d285f0e29abb994fe32db87ee81b00f403bd2]]
Former-commit-id: 120898e7b3979f399ed8eb0bd259d308366664f9
Former-commit-id: e895303771742980ac6b65488129cfc66bace5f5