rebase: use reflog to find common base with upstream
Commit 15a147e (rebase: use @{upstream} if no upstream specified,
2011-02-09) says:

	Make it default to 'git rebase @{upstream}'. That is also what
	'git pull [--rebase]' defaults to, so it only makes sense that
	'git rebase' defaults to the same thing.

but that isn't actually the case.  Since commit d44e712 (pull: support
rebased upstream + fetch + pull --rebase, 2009-07-19), pull has actually
chosen the most recent reflog entry which is an ancestor of the current
branch if it can find one.

Add a '--fork-point' argument to git-rebase that can be used to trigger
this behaviour.  This option is turned on by default if no non-option
arguments are specified on the command line, otherwise we treat an
upstream specified on the command-line literally.

Signed-off-by: John Keeping <john@keeping.me.uk>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: cb4f258e5bc737d4899313aaa6e51d8ad3bf4f32 [formerly a80a228b59fa8c33b6d728c86aeb4edcc57a3adf] [formerly a80a228b59fa8c33b6d728c86aeb4edcc57a3adf [formerly ad8261d21221d27638c75f47b39892db6f7972f6]]
Former-commit-id: dba7e7822ddb1abef042b53cfab3223eff4ba9a5
Former-commit-id: 8d7b0248b215721e06c1fa90805cae47d84be058