builtin/replace: teach listing using short, medium or full formats
By default when listing replace refs, only the sha1 of the
replaced objects are shown.

In many cases, it is much nicer to be able to list all the
sha1 of the replaced objects along with the sha1 of the
replacment objects.

And in other cases it might be interesting to also show the
types of the replaced and replacement objects.

This patch introduce a new --format=<fmt> option where
<fmt> can be any of the following:

	'short': this is the same as when no --format
		option is used, that is only the sha1 of
		the replaced objects are shown
	'medium': this also lists the sha1 of the
		replacement objects
	'full': this shows the sha1 and the type of both
		the replaced and the replacement objects

Some documentation and some tests will follow.

Signed-off-by: Christian Couder <chriscool@tuxfamily.org>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: c05f8a5909c503ac6a395c3e961941f2976f2bc0 [formerly 4291f0d3e38f6d8623e7045268ff8bf9aa7969e9] [formerly 4291f0d3e38f6d8623e7045268ff8bf9aa7969e9 [formerly 44f9f850e8ee6e2824ceb1855d836d484340edf7]]
Former-commit-id: 4a145917c019584223f9dfc3e5388887e9b04eda
Former-commit-id: c90420db4f81ad4184fefb235b658e0a57f94f57