builtin/replace: unset read_replace_refs
When checking to see if some objects are of the same type
and when displaying the type of objects, git replace uses
the sha1_object_info() function.

Unfortunately this function by default respects replace
refs, so instead of the type of a replaced object, it
gives the type of the replacement object which might
be different.

To fix this bug, and because git replace should work at a
level before replacement takes place, let's unset the
read_replace_refs global variable at the beginning of
cmd_replace().

Suggested-by: Jeff King <peff@peff.net>
Signed-off-by: Christian Couder <chriscool@tuxfamily.org>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: b33029048f9a5143957b8a5f8d7c34781cdb41a8 [formerly 36f6941dff71240b2465679ad80a3ed9bab75a42] [formerly 36f6941dff71240b2465679ad80a3ed9bab75a42 [formerly 769a4fa463bb36ba78eb318f25e0e211f4fae949]]
Former-commit-id: ccbd6b09fcdfea5879b5a744f1eea5bdb20840e5
Former-commit-id: 6e61f377c1b74d886dfbbcf999bfa21c193c0e25