diff: move no-index detection to builtin/diff.c
Currently the --no-index option is parsed in diff_no_index().  Move the
detection if a no-index diff should be executed to builtin/diff.c, where
we can use it for executing diff_no_index() conditionally.  This will
also allow us to execute other operations conditionally, which will be
done in the next patch.

There are no functional changes.

Helped-by: Jonathan Nieder <jrnieder@gmail.com>
Signed-off-by: Thomas Gummerer <t.gummerer@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 343868776658918bf70d992499427f98bf320338 [formerly 9c7541762baf5d1f8262157ba6711f5287033c8b] [formerly 9c7541762baf5d1f8262157ba6711f5287033c8b [formerly 470faf96544c1844ad775c695dfa370fe0ef3756]]
Former-commit-id: f8a36e6efbcd8c1e8566bac84bff973b75844143
Former-commit-id: 68929c4bbc5de41c06b861e7af6a10fca68f3628