diff: don't read index when --no-index is given
git diff --no-index ... currently reads the index, during setup, when
calling gitmodules_config().  This results in worse performance when the
index is not actually needed.  This patch avoids calling
gitmodules_config() when the --no-index option is given.  The times for
executing "git diff --no-index" in the WebKit repository are improved as
follows:

Test                      HEAD~3            HEAD
------------------------------------------------------------------
4001.1: diff --no-index   0.24(0.15+0.09)   0.01(0.00+0.00) -95.8%

An additional improvement of this patch is that "git diff --no-index" no
longer breaks when the index file is corrupt, which makes it possible to
use it for investigating the broken repository.

To improve the possible usage as investigation tool for broken
repositories, setup_git_directory_gently() is also not called when the
--no-index option is given.

Also add a test to guard against future breakages, and a performance
test to show the improvements.

Signed-off-by: Thomas Gummerer <t.gummerer@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 305bdd7e07a6acdb0bcfac9cb851c8319e74dbb9 [formerly c557ed4aa7488903d82e26a6193e1e8dafb5de7d] [formerly c557ed4aa7488903d82e26a6193e1e8dafb5de7d [formerly 6df5762db354ca55a0cf77451d06b332b7de0b82]]
Former-commit-id: 5a39719af5816be4816c9c858982afbdb737d782
Former-commit-id: 7f4d94b3c7a460ff9b7239e99ec869766f719266