Prevent buffer overflows when path is too long
Some buffers created with PATH_MAX length are not checked when being
written, and can overflow if PATH_MAX is not big enough to hold the
path.

Replace those buffers by strbufs so that their size is automatically
grown if necessary. They are created as static local variables to avoid
reallocating memory on each call. Note that prefix_filename() returns
this static buffer so each callers should copy or use the string
immediately (this is currently true).

Reported-by: Wataru Noguchi <wnoguchi.0727@gmail.com>
Signed-off-by: Antoine Pelisse <apelisse@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 79a355f7f6e096a8d087067d56bb090e92499f9e [formerly 6076241c98edbd728c5873eb2569e5378d92d499] [formerly 6076241c98edbd728c5873eb2569e5378d92d499 [formerly fc2b6214542a46f97d7067b2f7df530ed37737a7]]
Former-commit-id: dd9228763751494c453b1a3f21958098ff8ea77e
Former-commit-id: 6f519c4218890265c2bc022497c852bca4f1ae5a