log: properly handle decorations with chained tags
git log did not correctly handle decorations when a tag object referenced
another tag object that was no longer a ref, such as when the second tag was
deleted.  The commit would not be decorated correctly because parse_object had
not been called on the second tag and therefore its tagged field had not been
filled in, resulting in none of the tags being associated with the relevant
commit.

Call parse_object to fill in this field if it is absent so that the chain of
tags can be dereferenced and the commit can be properly decorated.  Include
tests as well to prevent future regressions.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 565b6e5a92b0b2725a0d543b88336ef827372f06 [formerly 565b6e5a92b0b2725a0d543b88336ef827372f06 [formerly 5e1361ccdbfaf328b5bae1eae29f16ba965c54b1]]
Former-commit-id: 9e3b45c56e4cea0f4f5fb7911bcc3a91070b4297
Former-commit-id: 80eb467ac77dbb19722655618dc4f0bc491e79d0