get_max_fd_limit(): fall back to OPEN_MAX upon getrlimit/sysconf failure
On broken systems where RLIMIT_NOFILE is visible by the compliers
but underlying getrlimit() system call does not behave, we used to
simply die() when we are trying to decide how many file descriptors
to allocate for keeping packfiles open.  Instead, allow the fallback
codepath to take over when we get such a failure from getrlimit().

The same issue exists with _SC_OPEN_MAX and sysconf(); restructure
the code in a similar way to prepare for a broken sysconf() as well.

Noticed-by: Joey Hess <joey@kitenet.net>
Helped-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: afb75a4981f3e9f242386abcbd01fb20c21c9fb7 [formerly afb75a4981f3e9f242386abcbd01fb20c21c9fb7 [formerly 491a8dec44e9b91149ef77c77c341e7d41df39be]]
Former-commit-id: a8031338e50be98dbeca26fa5ef8116186d63f84
Former-commit-id: 2ea19ae63449ad724dad61250dc720b8b9db3f0a