repack: handle optional files created by pack-objects
We ask pack-objects to pack to a set of temporary files, and
then rename them into place. Some files that pack-objects
creates may be optional (like a .bitmap file), in which case
we would not want to call rename(). We already call stat()
and make the chmod optional if the file cannot be accessed.
We could simply skip the rename step in this case, but that
would be a minor regression in noticing problems with
non-optional files (like the .pack and .idx files).

Instead, we can now annotate extensions as optional, and
skip them if they don't exist (and otherwise rely on
rename() to barf).

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 46ded9b86237b03792de35605ddec766b76e63a3 [formerly 24f1b29992fcaa40fa88f9a1728760d62a5acc8c] [formerly 24f1b29992fcaa40fa88f9a1728760d62a5acc8c [formerly b77fcd1edca58621627d46af0321a9737b1310f0]]
Former-commit-id: 60d98511a21d61820f75755c01238dff7e341241
Former-commit-id: e04f76df4e4b5b1a63174d6384d09792effb9cda