repack: consider bitmaps when performing repacks
Since `pack-objects` will write a `.bitmap` file next to the `.pack` and
`.idx` files, this commit teaches `git-repack` to consider the new
bitmap indexes (if they exist) when performing repack operations.

This implies moving old bitmap indexes out of the way if we are
repacking a repository that already has them, and moving the newly
generated bitmap indexes into the `objects/pack` directory, next to
their corresponding packfiles.

Since `git repack` is now capable of handling these `.bitmap` files,
a normal `git gc` run on a repository that has `pack.writebitmaps` set
to true in its config file will generate bitmap indexes as part of the
garbage collection process.

Alternatively, `git repack` can be called with the `-b` switch to
explicitly generate bitmap indexes if you are experimenting
and don't want them on all the time.

Signed-off-by: Vicent Marti <tanoku@gmail.com>
Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 7c8755de0acceb395983c0c94548211b13c766e2 [formerly e421f005db00d63232718ea879f1a58d8211a646] [formerly e421f005db00d63232718ea879f1a58d8211a646 [formerly 5cf2741c5a55cfb63076ce679e4b5842595125c2]]
Former-commit-id: a251b7fe1d13d5d51f9d53d2ceaa6181a9e53715
Former-commit-id: d3f19ace5a90ba856612ff8bb00eedf1f95d1128