t: add basic bitmap functionality tests
Now that we can read and write bitmaps, we can exercise them
with some basic functionality tests. These tests aren't
particularly useful for seeing the benefit, as the test
repo is too small for it to make a difference. However, we
can at least check that using bitmaps does not break anything.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: ac7ec176de049914f3b4b1e101bb8b5394ab6952 [formerly a5ea1d510f3d5ab4f6387f620cea6de1df30741e] [formerly a5ea1d510f3d5ab4f6387f620cea6de1df30741e [formerly 212f2ffbf0331ceca1a4bc1820ab1ac5c5115e8b]]
Former-commit-id: 0e79976972e54fa6fcd3ca34a02158944aed9e27
Former-commit-id: d311e50d4b1e0e4f8a0a85ffa471d5b23002a836