do not pretend sha1write returns errors
The sha1write function returns an int, but it will always be
"0". The failure-prone parts of the function happen in the
"flush" callback, which cannot pass an error back to us. So
we just end up calling die() during the flush.

Let's just drop the return value altogether, as it only
confuses callers into thinking that it might be useful.

Only one call site actually checked the return value. We can
drop that check, since it just led to a die() anyway.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: ef8a0828413bd8347ba890973879fd5243a2e620 [formerly 7766a81b24f424c685382a0aecb07780f07eb265] [formerly 7766a81b24f424c685382a0aecb07780f07eb265 [formerly 9af270e8c2a02afd9a3262a2c9312ee7fefbb7a3]]
Former-commit-id: d0f31d19b6ed994fe1a4b311ac8641044bbe531a
Former-commit-id: b231bbfbf634580d7eb120f4fed3d25ad003dc58