cat-file: provide %(deltabase) batch format
It can be useful for debugging or analysis to see which
objects are stored as delta bases on top of others. This
information is available by running `git verify-pack`, but
that is extremely expensive (and is harder than necessary to
parse).

Instead, let's make it available as a cat-file query format,
which makes it fast and simple to get the bases for a subset
of the objects.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: b4552d67d493e286edd44657e95c4d95e2400147 [formerly b4552d67d493e286edd44657e95c4d95e2400147 [formerly 65ea9c3c3d0e74b1f8c0c1d9fea6988550133dba]]
Former-commit-id: 6ec87b3f58f4678f0093c2f9792da07ed1c59a58
Former-commit-id: bb8ce8cdfa468f4aece4981102d3e5de22af4392