git-svn: workaround for a bug in svn serf backend
Subversion serf backend in versions 1.8.5 and below has a bug(*) that the
function creating the descriptor of a file change -- add_file() --
doesn't make a copy of its third argument when storing it on the
returned descriptor.  As a result, by the time this field is used (in
transactions of file copying or renaming) it may well be released, and
the memory reused.

One of its possible manifestations is the svn assertion triggering on an
invalid path, with a message

svn_fspath__skip_ancestor: Assertion
`svn_fspath__is_canonical(child_fspath)' failed.

This patch works around this bug, by storing the value to be passed as
the third argument to add_file() in a local variable with the same scope
as the file change descriptor, making sure their lifetime is the same.

* [ew: fixed in Subversion r1553376 as noted by Jonathan Nieder]

Cc: Benjamin Pabst <benjamin.pabst85@gmail.com>
Signed-off-by: Eric Wong <normalperson@yhbt.net>
Reviewed-by: Jonathan Nieder <jrnieder@gmail.com>
Signed-off-by: Roman Kagan <rkagan@mail.ru>

Former-commit-id: 1243c294487e7541394225c6e6c02de7890c4bf2 [formerly ca20c6e211105ad8cd2b114f68f972f937e9d42e] [formerly ca20c6e211105ad8cd2b114f68f972f937e9d42e [formerly ac930287ff16823531b541a8e6a7436279e612be]]
Former-commit-id: 750b8ff84827e2617ddc4cbac6245428129e3177
Former-commit-id: ec46cee2883abf9bca592e10c7d5ce892d293711