t0000: simplify HARNESS_ACTIVE hack
Commit 517cd55 set HARNESS_ACTIVE unconditionally in
sub-tests, because that value affects the output of
"--verbose". t0000 needs stable output from its sub-tests,
and we may or may not be running under a TAP harness.

That commit made the decision to always set the variable,
since it has another useful side effect, which is
suppressing writes to t/test-results by the sub-tests (which
would just pollute the real results).

Since the last commit, though, the sub-tests have their own
test-results directories, so this is no longer an issue. We
can now update a few comments that are no longer accurate
nor necessary.

We can also revisit the choice of HARNESS_ACTIVE. Since we
must choose one value for stability, it's probably saner to
have it off. This means that future patches could test
things like the test-results writing, or the "--quiet"
option, which is currently ignored when run under a harness.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 3c00254b81780ce74cc9313bff82ccc4d1e551f7 [formerly ac2232beda6d3a36514d232d8fea3d4720f0f6ef] [formerly ac2232beda6d3a36514d232d8fea3d4720f0f6ef [formerly a63c12c9be28bf455816b1ece4a52b463ecf6241]]
Former-commit-id: 2647c7cabf2df95c6c3b62477f340165ee71fbed
Former-commit-id: 60fa697d36d9b79e4b3f90531d65f5e78007dcb9