t0000: drop "known breakage" test
Having a simulated "known breakage" test means that the test
suite will always tell us there is a bug to be fixed, even
though it is only simulated.

The right way to test this is in a sub-test, that can also
check that we provide the correct exit status and output.
Fortunately, we already have such a test (added much later
by 5ebf89e).

We could arguably get rid of the simulated success test
immediately above, as well, as it is also redundant with the
tests added in 5ebf89e. However, it does not have the
annoying behavior of the "known breakage" test. It may also
be easier to debug if the test suite is truly broken, since
it is not a test-within-a-test, as the later tests are.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: de79caefa7f58173bb267a8a325385a8848fd38d [formerly de79caefa7f58173bb267a8a325385a8848fd38d [formerly 738a8beac42d5e2c6b882f997b7fc6577363c544]]
Former-commit-id: e07ca031380d8b2d8e962eb87b87a7e640405f50
Former-commit-id: e94f5ef821088109378b86ca4cdad7a66058ad03