gc: notice gc processes run by other users
Since 64a99eb4 git gc refuses to run without the --force option if
another gc process on the same repository is already running.

However, if the repository is shared and user A runs git gc on the
repository and while that gc is still running user B runs git gc on
the same repository the gc process run by user A will not be noticed
and the gc run by user B will go ahead and run.

The problem is that the kill(pid, 0) test fails with an EPERM error
since user B is not allowed to signal processes owned by user A
(unless user B is root).

Update the test to recognize an EPERM error as meaning the process
exists and another gc should not be run (unless --force is given).

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 03df5b05726d2b563a2fb4e75645c1aa20b9b6f0 [formerly 6a5117600ed28dd4433425a6234c4488cad94918] [formerly 6a5117600ed28dd4433425a6234c4488cad94918 [formerly ed7eda8b38ec6230abf79986668ccb6e2e2c7494]]
Former-commit-id: 49f1cec9c300726e101c408d24a021d1b7aceb86
Former-commit-id: 070585b7e1e8d5fa73c9280584b34ba5ca0cb078