fetch --prune: Run prune before fetching
When we have a remote-tracking branch named "frotz/nitfol" from a
previous fetch, and the upstream now has a branch named "frotz",
fetch would fail to remove "frotz/nitfol" with a "git fetch --prune"
from the upstream. git would inform the user to use "git remote
prune" to fix the problem.

Change the way "fetch --prune" works by moving the pruning operation
before the fetching operation. This way, instead of warning the user
of a conflict, it autmatically fixes it.

Signed-off-by: Tom Miller <jackerran@gmail.com>
Tested-by: Thomas Rast <tr@thomasrast.ch>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 480102326cb29f36b7394ff3910b0ad4336babf6 [formerly e9b8e35a2c4135b7bd3b005ddf87916c25340309] [formerly e9b8e35a2c4135b7bd3b005ddf87916c25340309 [formerly 10a6cc8890ec1e5459c05ddeb28a671acdc37d60]]
Former-commit-id: 6e9fb0b21f87b16ea5d92d201c69d05eabda766d
Former-commit-id: 455a43c37b90a1c7125f27744a95052ace1df734