Documentation/gitmodules: Only 'update' and 'url' are required
Descriptions for all the settings fell under the initial "Each
submodule section also contains the following required keys:".  The
example shows sections with just 'path' and 'url' entries, which are
indeed required, but we should still make the required/optional
distinction explicit to clarify that the rest of them are optional.

Signed-off-by: W. Trevor King <wking@tremily.us>
Reviewed-by: Heiko Voigt <hvoigt@hvoigt.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 6e7db517986760cec60b2ab7e59003ac9e0318b4 [formerly 6e7db517986760cec60b2ab7e59003ac9e0318b4 [formerly 43fda9455c2f9300327e5cf1c532c083354e66b5]]
Former-commit-id: aae849f262cf83aa9be304d212fa38eb5056da0a
Former-commit-id: 018d68815ba9b47ad8ff2bd971bf6b169c497fea