git-submodule.sh: 'checkout' is a valid update mode
'checkout' is documented as one of the valid values for the
'submodule.<name>.update' variable, and in a repository with the
variable set to 'checkout', "git submodule update" command does
update using the 'checkout' mode.

However, it has been an accident that the implementation works this
way; any unknown value would trigger the same codepath and update
using the 'checkout' mode.

Explicitly list 'checkout' as one of the known update modes, and
error out when an unknown update mode is used.

Teach the codepath that initializes the configuration variable from
an in-tree .gitmodules that 'checkout' is one of the valid values.
The code since ac1fbbda (submodule: do not copy unknown update mode
from .gitmodules, 2013-12-02) used to treat the value 'checkout' as
unknown and mapped it to 'none', which made little sense.  With this
change, 'checkout' specified in .gitmodules will stay to be 'checkout'.

Signed-off-by: Francesco Pretto <ceztko@gmail.com>
Signed-off-by: Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 4d03989ae8c8f7ff064f10f3c34f467f5d65a8e8 [formerly 4d03989ae8c8f7ff064f10f3c34f467f5d65a8e8 [formerly efa8fd7ee8eb36df55b3549f6009f000347a796d]]
Former-commit-id: d7cf5c9b770c3f00e001f2633ae7679ffb65379e
Former-commit-id: c22e54809a10f551300afc16a7bf54d9d3f7f564