zsh completion: find matching custom bash completion
If zsh completion is being read from a location that is different from
system-wide default, it is likely that the user is trying to use a
custom version, perhaps closer to the bleeding edge, installed in her
own directory. We will more likely to find the matching bash completion
script in the same directory than in those system default places.

Signed-off-by: Ramkumar Ramachandra <artagnon@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: ee1bc5841acd95de5bee41f8ae34dada1837b7b5 [formerly ee1bc5841acd95de5bee41f8ae34dada1837b7b5 [formerly d028b8906afb5b8c24c9449f26e5f930750d8df7]]
Former-commit-id: 76aa492e9cfde219ba8f47536e3abe254924432e
Former-commit-id: 2a7945dadfa664431817ac6a4d1f71859662f995