completion: introduce __gitcomp_nl_append ()
There are situations where multiple classes of completions possible. For
example

  branch.<TAB>

should try to complete

  branch.master.
  branch.autosetupmerge
  branch.autosetuprebase

The first candidate has the suffix ".", and the second/ third candidates
have the suffix " ". To facilitate completions of this kind, create a
variation of __gitcomp_nl () that appends to the existing list of
completion candidates, COMPREPLY.

Signed-off-by: Ramkumar Ramachandra <artagnon@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 61b6ea14616bd072321f8296e10e16c5521bfdac [formerly 9744fd59b23cd0fc89ea9c73dcde4afae99b6b0c] [formerly 9744fd59b23cd0fc89ea9c73dcde4afae99b6b0c [formerly f33c2c0f9e2258a80e6e406af39e9fa3fd6e430e]]
Former-commit-id: ff602f3e795e84b3d5a681829a1c6c80c6904fb4
Former-commit-id: 067de96f6bac17db337721d9093fab3b3a51535d