safe_create_leading_directories(): split on first of multiple slashes
If the input path has multiple slashes between path components (e.g.,
"foo//bar"), then the old code was breaking the path at the last
slash, not the first one.  So in the above example, the second slash
was overwritten with NUL, resulting in the parent directory being
sought as "foo/".

When stat() is called on "foo/", it fails with ENOTDIR if "foo" exists
but is not a directory.  This caused the wrong path to be taken in the
subsequent logic.

So instead, split path components at the first intercomponent slash
rather than the last one.

Signed-off-by: Michael Haggerty <mhagger@alum.mit.edu>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 935213ffab6b3400ab9a0eaa170dc78b77011f2e [formerly 24db0fa91b7ab4a63b7a4df1b217128a28301e7b] [formerly 24db0fa91b7ab4a63b7a4df1b217128a28301e7b [formerly bf10cf70ad0c777dbbbb00bbb741436e285c2181]]
Former-commit-id: 0a1bc0365c300d5144358856b6d6c6b29812adff
Former-commit-id: 5b8124524be23816a1b3b2768cedc0faf5dd8f2b