safe_create_leading_directories(): add new error value SCLD_VANISHED
Add a new possible error result that can be returned by
safe_create_leading_directories() and
safe_create_leading_directories_const(): SCLD_VANISHED.  This value
indicates that a file or directory on the path existed at one point
(either it already existed or the function created it), but then it
disappeared.  This probably indicates that another process deleted the
directory while we were working.  If SCLD_VANISHED is returned, the
caller might want to retry the function call, as there is a chance
that a new attempt will succeed.

Why doesn't safe_create_leading_directories() do the retrying
internally?  Because an empty directory isn't really ever safe until
it holds a file.  So even if safe_create_leading_directories() were
absolutely sure that the directory existed before it returned, there
would be no guarantee that the directory still existed when the caller
tried to write something in it.

Signed-off-by: Michael Haggerty <mhagger@alum.mit.edu>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 2b308dac98b32c92fe874ce1e6092bbdaf7c0762 [formerly 2b308dac98b32c92fe874ce1e6092bbdaf7c0762 [formerly 18d37e860dfb9a98fb93ea7bb517ec3c16f995c4]]
Former-commit-id: 6a2f96ad935909f3d792c7eef2e4a4a2c0718c59
Former-commit-id: e559d56a9b15d88571ef5064efa377155b3421fd