sha1_name: don't resolve refs when core.warnambiguousrefs is false
When seeing a full 40-hex object name, get_sha1_basic()
unconditionally checks if the string can also be interpreted as a
refname, but the result will not be used unless warn_ambiguous_refs
is in effect.

Omitting this unnecessary ref resolution provides a substantial
performance improvement, especially when passing many hashes to a
command (like "git rev-list --stdin") and core.warnambiguousrefs is
set to false.  The check incurs 6 stat()s for every hash supplied,
which can be costly over NFS.

Signed-off-by: Brodie Rao <brodie@sf.io>
Acked-by: Nguyễn Thái Ngọc Duy <pclouds@gmail.com>
Acked-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 8411f994d591c9eb5c56a530d763e89ffdd15337 [formerly dc448ea754a6fe2acf8aac88a8d6af5579b0bf8f] [formerly dc448ea754a6fe2acf8aac88a8d6af5579b0bf8f [formerly 832cf74c0792a58c9c28e32a8fe5dbb694f0cce6]]
Former-commit-id: fa90a89a815a1c093ddf088ce2ea4d3ba684c505
Former-commit-id: 624c838d48453b540a2d49fc61a4e6dd8c62e553