stash: handle specifying stashes with $IFS
When trying to pop/apply a stash specified with an argument
containing IFS whitespace, git-stash will throw an error:

    $ git stash pop 'stash@{two hours ago}'
    Too many revisions specified: stash@{two hours ago}

This happens because word splitting is used to count non-option
arguments. Make use of rev-parse's --sq option to quote the arguments
for us to ensure a correct count. Add quotes where necessary.

Also add a test that verifies correct behaviour.

Helped-by: Thomas Rast <tr@thomasrast.ch>
Signed-off-by: Øystein Walle <oystwa@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 853ee8f827fbe7a848a0c1de9a172c648b1fcb15 [formerly 853ee8f827fbe7a848a0c1de9a172c648b1fcb15 [formerly 2a07e4374c0ba6f2e991965c99b448ccb563f2fc]]
Former-commit-id: 1cf6766f7b7cf054c5f5e9fc8fbc0550dd9aea9b
Former-commit-id: ebefe631fc5870b45327feeb7143d7dff86a46a5