mv: better document side effects when moving a submodule
The "Submodules" section of the "git mv" documentation mentions what will
happen when a submodule with a gitfile gets moved with newer git. But it
doesn't talk about what happens when the user changes between commits
before and after the move, which does not update the work tree like using
the mv command did the first time.

Explain what happens and what the user has to do manually to fix that in
the new BUGS section. Also document this behavior in a new test.

Reported-by: George Papanikolaou <g3orge.app@gmail.com>
Signed-off-by: Jens Lehmann <Jens.Lehmann@web.de>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 8e47d7ea80a6532a6a465da6cedaeb420d41077b [formerly a5fccf571d3a0cebc10c88e8a944b38476bc27c4] [formerly a5fccf571d3a0cebc10c88e8a944b38476bc27c4 [formerly 1cbd18300a8755ba46791b2aa6249fa537a1d651]]
Former-commit-id: 88b30b9c03c67a5b5bedab8b9126df090923e0d3
Former-commit-id: 921343a89265afb636092254eb6768a6ea55cd6d