rm: better document side effects when removing a submodule
The "Submodules" section of the "git rm" documentation mentions what will
happen when a submodule with a gitfile gets removed with newer git. But it
doesn't talk about what happens when the user changes between commits
before and after the removal, which does not remove the submodule from the
work tree like using the rm command did the first time.

Explain what happens and what the user has to do manually to fix that in
the new BUGS section. Also document this behavior in a new test.

Signed-off-by: Jens Lehmann <Jens.Lehmann@web.de>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: bb3b102b52be3d1132f43265ea74d1a240fe3b4f [formerly bb3b102b52be3d1132f43265ea74d1a240fe3b4f [formerly bbad9f9314f658b5c0f302148fc9780f5788dcd8]]
Former-commit-id: 741925cc11009f30eed288e61acbc11dfffaab4c
Former-commit-id: 8c4b6d1845db0974dcfc8ae4022ca4efaa8c8204