t5531: further "matching" fixups
Commit 43eb920 switched one of the sub-repository in this
test to matching to prepare for a world where the default
becomes "simple". However, the main repository needs a
similar change.

We did not notice any test failure when merged with b2ed944
(push: switch default from "matching" to "simple", 2013-01-04)
because t5531.6 is trying to provoke a failure of "git push"
due to a submodule check. When combined with b2ed944 the
push still fails, but for the wrong reason (because our
upstream setup does not exist, not because of the submodule).

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 9b1b6ffc1b4b973593697ce3698c5f49aab7328d [formerly 0af34ce8056f4f8b473bd4e03e17af3241ff4f25] [formerly 0af34ce8056f4f8b473bd4e03e17af3241ff4f25 [formerly 0b1985050e512d6e0a5fff084ecc7c2232596065]]
Former-commit-id: 545d7258885c04db5cc7e97fc7209e537ade3a90
Former-commit-id: 28f5f2ab8e982799b672356f69e49b7d9e720403