t5537: fix incorrect expectation in test case 10
Commit 48d25ca adds a new commit "7" to the repo that the next test case
in commit 1609488 clones from. But the next test case does not expect
this commit. For these tests, it's the bottom that's important, not
the top. Fix the expected commit list.

While at it, fix the default http port number to 5537. Otherwise when
t5536 learns to test httpd, running test in parallel may fail.

References:

48d25ca fetch: add --update-shallow to accept... - 2013-12-05
1609488 smart-http: support shallow fetch/clone - 2013-12-05

Noticed-by: Jeff King <peff@peff.net>
Signed-off-by: Nguyễn Thái Ngọc Duy <pclouds@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: e492a949dadb40cdc9f23795acb36ee38772618e [formerly e492a949dadb40cdc9f23795acb36ee38772618e [formerly 3b32a7ca90b9c63f2306feb2a66b62b94c1a640f]]
Former-commit-id: 14e60992f161818ef5ea0dd9cec9e01492a910f0
Former-commit-id: 65da2ce26c7b5b53246c0bec6e47e3eedce53fee