mv: let 'git mv file no-such-dir/' error out on Windows, too
The previous commit c57f628 (mv: let 'git mv file no-such-dir/' error out)
relies on that rename("file", "no-such-dir/") fails if the directory does not
exist (note the trailing slash).  This does not work as expected on Windows:
This rename() call does not fail, but renames "file" to "no-such-dir" (not to
"no-such-dir/file"). Insert an explicit check for this case to force an error.

This changes the error message from

   $ git mv file no-such-dir/
   fatal: renaming 'file' failed: Not a directory

to

   $ git mv file no-such-dir/
   fatal: destination directory does not exist, source=file, destination=no-such-dir/

Signed-off-by: Johannes Sixt <j6t@kdbg.org>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: b8bb68cfc7f6e06d571347a990e961abb272374f [formerly 3d638871ff709768dbd197aa5bb34e45e1af3ec8] [formerly 3d638871ff709768dbd197aa5bb34e45e1af3ec8 [formerly a8933469309c492ad69af3f25bfddc7b245ab9c3]]
Former-commit-id: 413ef20e7a4f26cf4f8d2d910530c39e561b052b
Former-commit-id: 8763f624ec922710ef52e1906a3b04ef3e65392f