rebase: fix fork-point with zero arguments
When no arguments are specified, $switch_to is empty so we end up
passing the empty string to "git merge-base --fork-point", which causes
an error.  git-rebase carries on at this point, but in fact we have
failed to apply the fork-point operation.

It turns out that the test in t3400 that was meant to test this didn't
actually need the fork-point behaviour, so enhance it to make sure that
the fork-point is applied correctly.  The modified test fails without
the change to git-rebase.sh in this patch.

Reported-by: Andreas Krey <a.krey@gmx.de>
Signed-off-by: John Keeping <john@keeping.me.uk>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 4f835e79224dcaa12f2e7d0c8635bd4804508224 [formerly 4f835e79224dcaa12f2e7d0c8635bd4804508224 [formerly bb3f45838b859c8b17a53a24579a304333868cc8]]
Former-commit-id: 12ef30cfee5d94e6ddcee2c22c81e87a7c214cee
Former-commit-id: 004145ea8cc82022c36a59d37e72b8f03f47e1cd