Documentation: "git pull" does not have the "-m" option
Even though "--[no-]edit" can be used with "git pull", the
explanation of the interaction between this option and the "-m"
option does not make sense within the context of "git pull".  Use
the conditional inclusion mechanism to remove this part from "git
pull" documentation, while keeping it for "git merge".

Reported-by: Ivan Zakharyaschev
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 661e0aa972af944c59eb26e399081d0d41fa99be [formerly 6dd88380278c6dca276f19412dd594e7fc5ce88b] [formerly 6dd88380278c6dca276f19412dd594e7fc5ce88b [formerly 08f19cfe9bff097e8828a289d2daf5f066e1af9d]]
Former-commit-id: cc7d173e493747d9b4dc0f82fe272b0eeaf1c3eb
Former-commit-id: d6120d23551d444673b28670375d5ba33946c084