git-gui: fallback right pane to packed widgets with Tk 8.4
Since 918dbf58, git-gui crashes if started with Tk 8.4. The reason is that
tk < 8.5 does not support -stretch option for panedwindow.

Without the option it's not possible to properly expand the right half -
the commit area is expanded, while desired behavior is to expand the diff
area. So the whole feature should be disabled with Tk
version less than 8.5.

Signed-off-by: Max Kirillov <max@max630.net>
Signed-off-by: Pat Thoyts <patthoyts@users.sourceforge.net>

Former-commit-id: 72bcb98d77ab70d48c06fa411604448d6e883c4a [formerly 72bcb98d77ab70d48c06fa411604448d6e883c4a [formerly 02f6cfbd1628d9e99a241d18d805f3842dc746e7]]
Former-commit-id: 9e5198544b5defe8c7d27645ca418e3ba31fdb69
Former-commit-id: 56c65e0654c4a32889bc60113b528c4acc8546ae