revision: propagate flag bits from tags to pointees
With the previous fix 895c5ba3 (revision: do not peel tags used in
range notation, 2013-09-19), handle_revision_arg() that processes
command line arguments for the "git log" family of commands no
longer directly places the object pointed by the tag in the pending
object array when it sees a tag object.  We used to place pointee
there after copying the flag bits like UNINTERESTING and
SYMMETRIC_LEFT.

This change meant that any flag that is relevant to later history
traversal must now be propagated to the pointed objects (most often
these are commits) while starting the traversal, which is partly
done by handle_commit() that is called from prepare_revision_walk().
We did propagate UNINTERESTING, but did not do so for others, most
notably SYMMETRIC_LEFT.  This caused "git log --left-right v1.0..."
(where "v1.0" is a tag) to start losing the "leftness" from the
commit the tag points at.

Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: e6e488bffa4db47b2c2ec34864e27ae6112b8930 [formerly 0556cb4f9c6e851cccd1325868b955806b9a2e89] [formerly 0556cb4f9c6e851cccd1325868b955806b9a2e89 [formerly a74352867e689d50ee9c368f24d4a64392e27a35]]
Former-commit-id: c4a2163ca1e6673e263d2872300d53fae10aba91
Former-commit-id: 73a1315700a9a2641353199b9b9b2a3ec5a216c2