diff_filespec: reorder dirty_submodule macro definitions
diff_filespec has a 2-bit "dirty_submodule" field and
defines two flags as macros. Originally these were right
next to each other, but a new field was accidentally added
in between in commit 4682d85. This patch puts the field and
its flags back together.

Using an enum like:

  enum {
	  DIRTY_SUBMODULE_UNTRACKED = 1,
	  DIRTY_SUBMODULE_MODIFIED = 2
  } dirty_submodule;

would be more obvious, but it bloats the structure. Limiting
the enum size like:

  } dirty_submodule : 2;

might work, but it is not portable.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 52e478b5d9a521b92ffd1abfdf42e2dc8654969d [formerly 91093d6ba55ee60d259cc0a18897703c73173c02] [formerly 91093d6ba55ee60d259cc0a18897703c73173c02 [formerly b837f5d68da391f3a54817bb04f1e7ec04f9cdac]]
Former-commit-id: 8f883ad3d10b9d55eef51cae620f3ed411078117
Former-commit-id: dbc47c7db3b933cfee31c360e62ef7e4bcd58f7a