diff_filespec: reorder is_binary field
The middle of the diff_filespec struct contains a mixture of
ints, shorts, and bit-fields, followed by a pointer. On an
x86-64 system with an LP64 or LLP64 data model (i.e., most
of them), the integers and flags end up being padded out by
41 bits to put the pointer at an 8-byte boundary.

After the pointer, we have the "int is_binary" field, which
is only 32 bits. We end up wasting another 32 bits to pad
the struct size up to a multiple of 64 bits.

We can move the is_binary field before the pointer, which
lets the compiler store it where we used to have padding.
This shrinks the top padding to only 9 bits (from the
bit-fields), and eliminates the bottom padding entirely,
dropping the struct size from 88 to 80 bytes.

On a 32-bit system, there is no benefit, but nor should
there be any harm (we only need 4-byte alignment there, so
we were already using only 9 bits of padding).

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 4e00a89000d283cd38937431b2be7d1b8894381a [formerly 7be0bfe1483b0552a9d2cda23200050e25be44e0] [formerly 7be0bfe1483b0552a9d2cda23200050e25be44e0 [formerly b38f70a82b9628d80f5a9bc938a8861ee4dcf49e]]
Former-commit-id: 783d43aa1d957c63439b75d516afbf3dc068e7dd
Former-commit-id: ba6e213ef3ba1a8f8c21aa93a599c16693c0566b