diff_filespec: use only 2 bits for is_binary flag
The is_binary flag needs only three values: -1, 0, and 1.
However, we use a whole 32-bit int for it on most systems
(both 32- and 64- bit).

Instead, we can mark it to use only 2 bits. On 32-bit
systems, this lets it end up as part of the bitfield above
(saving 4 bytes). On 64-bit systems, we don't see any change
(because the savings end up as padding), but it does leave
room for another "free" 32-bit value to be added later.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: da3de571364ff4d3d9fd5d2f0779c4b4b364a577 [formerly da3de571364ff4d3d9fd5d2f0779c4b4b364a577 [formerly cbfe47b67fc1072998c73e6d43cf6ad061a436f5]]
Former-commit-id: 88ce193263bd382a77a23a5a38b56aef6a144ed7
Former-commit-id: 72ebe98bca3f144a6a43a52cc3ff35d615db2e35