pull: suppress error when no remoteref is found
Commit 48059e4 (pull: use merge-base --fork-point when appropriate,
2013-12-08) incorrectly assumes that get_remote_merge_branch will either
yield a non-empty string or return an error, but there are circumstances
where it will yield an empty string.

The previous code then invoked git-rev-list with no arguments, which
results in an error suppressed by redirecting stderr to /dev/null.  Now
we invoke git-merge-base with an empty branch name, which also results
in an error.  Suppress this in the same way.

Signed-off-by: John Keeping <john@keeping.me.uk>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 0cb2ddd44b9b2a62df910b2e3f26cc97e2d49735 [formerly 12856be2537538cd174b99f5e6a8596b893405ad] [formerly 12856be2537538cd174b99f5e6a8596b893405ad [formerly ffc2b483dee43c2687088d4c73d7532fda855050]]
Former-commit-id: 98a897ecbe5ee2a483412ae84ee346217a177b19
Former-commit-id: 4ab18ae5a993ca8db71dc400a25e3a5120a4328b