lock_ref_sha1_basic(): on SCLD_VANISHED, retry
If safe_create_leading_directories() fails because a file along the
path unexpectedly vanished, try again (up to 3 times).

This can occur if another process is deleting directories at the same
time as we are trying to make them.  For example, "git pack-refs
--all" tries to delete the loose refs and any empty directories that
are left behind.  If a pack-refs process is running, then it might
delete a directory that we need to put a new loose reference in.

If safe_create_leading_directories() thinks this might have happened,
then take its advice and try again (maximum three attempts).

Signed-off-by: Michael Haggerty <mhagger@alum.mit.edu>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 2febc38c0c466be49e3d7d24e877ff0570e14904 [formerly 2febc38c0c466be49e3d7d24e877ff0570e14904 [formerly c4c61c763e700d02344490590d6980ee51031a27]]
Former-commit-id: a312bbb2d10aff1c5509630da1dbce701d7c89ed
Former-commit-id: 0f5dcc932a3907475cc3b05b501dc45e42375cce