rename_tmp_log(): limit the number of remote_empty_directories() attempts
This doesn't seem to be a likely error, but we've got the counter
anyway, so we might as well use it for an added bit of safety.

Please note that the first call to rename() is optimistic, and it is
normal for it to fail if there is a directory in the way.  So bump the
total number of allowed attempts to 4, to be sure that we can still
have at least 3 retries in the case of a race.

Signed-off-by: Michael Haggerty <mhagger@alum.mit.edu>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: e9e49abdf8f183a296a6cd922c84dc0fda641892 [formerly e9e49abdf8f183a296a6cd922c84dc0fda641892 [formerly f1e9e9a4dbe2cfb39dcb14ee4f34628ef46d7b15]]
Former-commit-id: e316c8d3ad29fb6ea32831e36306c1ff7067cc0a
Former-commit-id: 7eed6087a28cb7486a12ee8308fa08e14b6241d7