combine-diff: combine_diff_path.len is not needed anymore
The field was used in order to speed-up name comparison and also to
mark removed paths by setting it to 0.

Because the updated code does significantly less strcmp and also
just removes paths from the list and free right after we know a path
will not be needed, it is not needed anymore.

Signed-off-by: Kirill Smelkov <kirr@mns.spb.ru>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 8590d84288216a63e5a124f28122409bbfe86311 [formerly 20bf85374698976b90b2c956cd9bb21b5ee23b46] [formerly 20bf85374698976b90b2c956cd9bb21b5ee23b46 [formerly af82c7880f1a3df1655092da11c80603260384a0]]
Former-commit-id: c22dff2e66104ad8331e6caec3a257e29efb7f3f
Former-commit-id: 7c7deab63f2f59a8600fcdec89c1840bdf286037