gitk: Comply with XDG base directory specification
Write the gitk config data to $XDG_CONFIG_HOME/git/gitk ($HOME/.config/git/gitk
by default) in line with the XDG specification. This makes it consistent with
git which also follows the spec.

If $HOME/.gitk already exists use that for backward compatibility, so only new
installations are affected.

Signed-off-by: Astril Hayato <astrilhayato@gmail.com>
Signed-off-by: Paul Mackerras <paulus@samba.org>

Former-commit-id: 7076957d2b56387596f852faf439f2218b63fc76 [formerly 7076957d2b56387596f852faf439f2218b63fc76 [formerly 8f86339858c1325645f8eaf6950dfd9d74cc9897]]
Former-commit-id: 79d5cd7329802d7030516e9aeae16f0500a0d5cc
Former-commit-id: 7afdcaed0b946c1c362ca717e556e5b046525cf7