git p4 test: ensure p4 symlink parsing works
While this happens to work, there was no test to make sure
that the basic importing of a symlink from p4 to git functioned.

Add a simple test to create a symlink in p4 and import it into git,
then verify that the symlink exists and has the correct target.

Signed-off-by: Pete Wyckoff <pw@padd.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 9c5557120c9f156948d6a37fa896bf938f7ecfbd [formerly 9c5557120c9f156948d6a37fa896bf938f7ecfbd [formerly a8d8e382a95b8412b105ca12fcdad2dff521827f]]
Former-commit-id: cc038ea7cba12f9e9b8624bff8ac0f1c9acccda6
Former-commit-id: 778ac8f38f0d0151d4ba6ba8a62d6808ec620708