git p4 test: run as user "author"
The tests use author@example.com as the canonical submitter,
but he does not have an entry in the p4 users database.
This causes the generated change description to complain
that the git and p4 users disagree.  The complaint message
is still valid, but isn't useful in tests.  It was introduced
in 848de9c (git-p4: warn if git authorship won't be retained,
2011-05-13).

Fix t9813 to use @example.com instead of @localhost due to
change in p4_add_user().  Move the function into the git p4
test library so author can be added at initialization time.

Signed-off-by: Pete Wyckoff <pw@padd.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 498883a32fa55eafdb799b35a0c956db2fd083fc [formerly 498883a32fa55eafdb799b35a0c956db2fd083fc [formerly 0055b56e10f77b1b5f5521b8ba3614f962a09288]]
Former-commit-id: 19816b7aeb373ff7b84c3479c663cd5f749761b4
Former-commit-id: ccf8ca2d3be615ed37967922b0076f94e8acbe2d