git p4 test: do not pollute /tmp
Generating the submit template for p4 uses tempfile.mkstemp(),
which by default puts files in /tmp.  For a test that fails,
possibly on purpose, this is not cleaned up.  Run with TMPDIR
pointing into the trash directory so the temp files go away
with the test results.

To do this required some other minor changes.  First, the editor
is launched using system(editor + " " + template_file), using
shell expansion to build the command string.  This doesn't work
if editor has a space in it.  And is generally unwise as it's
easy to fool the shell into doing extra work.  Exec the args
directly, without shell expansion.

Second, without shell expansion, the trick of "P4EDITOR=:" used
in the tests doesn't work.  Use a real command, true, as the
non-interactive editor for testing.

Signed-off-by: Pete Wyckoff <pw@padd.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 1250c45813c835d21d6ee9d51d7ac2265a24344f [formerly bf62791f45224dab66da48b63047a6ad9411e217] [formerly bf62791f45224dab66da48b63047a6ad9411e217 [formerly 0cf1b72a38e6190a7e614bbc53fbb81704a3d4af]]
Former-commit-id: 3f468017db28bc51888aeb0e39d111b473cf58fe
Former-commit-id: 4529a3ebecc23737711887246b2861f7ea11cc08