git p4 test: examine behavior with locked (+l) files
The p4 server can enforce file locking, so that only one user
can edit a file at a time.  Git p4 is unable to submit changes
to locked files.  Currently it exits poorly.  Ideally it would
notice the locked condition and clean up nicely.

Add a bunch of tests that describe the problem, hoping that
fixes appear in the future.

Signed-off-by: Pete Wyckoff <pw@padd.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: ec4465d117862b858c9d8cebe80bd5d22dd07fda [formerly 1104f0b56d0728602e00cde56c1757f68be7b845] [formerly 1104f0b56d0728602e00cde56c1757f68be7b845 [formerly 3d5388afa856fde3efff023f5b672005834e7ca1]]
Former-commit-id: d62bf296615421a34d4f0f26f27f2bed1dc2325e
Former-commit-id: 0aafdcf13a27b3bd3a64292bf6a605651066d6ad