t7501: fix "empty commit" test with NO_PERL
t7501.9 tries to check that "git commit" will fail when the
index is unchanged. It relies on previous tests not to have
modified the index. When it was originally written, this was
always the case. However, commit c65dc35 (t7501: test the
right kind of breakage, 2012-03-30) changed earlier tests (4
and 5) to leave a modification in the index.

We never noticed, however, because t7501.7, between the two,
clears the index state as a side effect. However, that test
depends on the PERL prerequisite, and so it does not always
run. Therefore if NO_PERL is set, we do not run the
intervening test, the index is left unclean, and t7501.9
fails.

We could fix this by moving t7501.9 up in the script.
However, this patch instead leaves it in place and adds a
"git reset" before the commit. This makes the test more
explicit about its preconditions, and will future-proof it
against any other changes in the test state.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: c0dc37a90df701a6e197575599b67fc09d5e26ba [formerly b3e94b9919e5ff2deb4a3eef10788ce8715bedbf] [formerly b3e94b9919e5ff2deb4a3eef10788ce8715bedbf [formerly 088304bf73b9b4149e04d2246fe08a06eef6e795]]
Former-commit-id: d53a4ebc5f65fab3b2259c106610f320197c842c
Former-commit-id: 7889d6ba53f4da67af86bd4063495df754d57b9c