t7700: do not use "touch" unnecessarily
Some versions of touch (such as /usr/ucb/touch on Solaris)
do not know about the "-r" option. This would make sense as
a feature of test-chmtime, but fortunately this fix is even
easier.

The test does not care about the timestamp of the .keep file it
creates at all, only that it exists. For such a use case, with or
without portability issues around "-r", "touch" should not be used
in the first place.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 2035468c0949fe0f587c80ca82578a5132b70f45 [formerly 2035468c0949fe0f587c80ca82578a5132b70f45 [formerly a0332337be53f266682279c72a5e553986638c87]]
Former-commit-id: 28be6632aa7d2513339a3d766d8a859f575dda7c
Former-commit-id: 02b30bfaca3e6e7cafecf3424f55cf15081b0a37