ewah: support platforms that require aligned reads
The caller may hand us an unaligned buffer (e.g., because it
is an mmap of a file with many ewah bitmaps). On some
platforms (like SPARC) this can cause a bus error. We can
fix it with a combination of get_be32 and moving the data
into an aligned buffer (which we would do anyway, but we can
move it before fixing the endianness).

Signed-off-by: Vicent Marti <tanoku@gmail.com>
Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 994833adeb3a998e8420132db9d4eccc6fc81bd4 [formerly 994833adeb3a998e8420132db9d4eccc6fc81bd4 [formerly a201c20b41a2f0725977bcb89a2a66135d776ba2]]
Former-commit-id: ba56589f6d644e5407146ecbaf2c33442686ad14
Former-commit-id: 75af9abc138ddfc4e2a0323855642b774eca0740