match_pathspec: match pathspec "foo/" against directory "foo"
Currently we do support matching pathspec "foo/" against directory
"foo". That is because match_pathspec() has no way to tell "foo" is a
directory and matching "foo/" against _file_ "foo" is wrong.

The callers can now tell match_pathspec if "foo" is a directory, we
could make an exception for this case. Code is not executed though
because no callers pass the flag yet.

Signed-off-by: Nguyễn Thái Ngọc Duy <pclouds@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: ec0d35d2e18d404749b32852ebcd860ce9da0cde [formerly fd6dece396809743ab98040257c5e7ac99ec0fde] [formerly fd6dece396809743ab98040257c5e7ac99ec0fde [formerly 68690fdd0b78762eb6387d7a437b588d15b6cf47]]
Former-commit-id: 4344627e7e28c43d1b337ba4a406238fb247b227
Former-commit-id: 798d7b46161eaf180f3390035ff66b3fa9ecbabf