pathspec: pass directory indicator to match_pathspec_item()
This patch activates the DO_MATCH_DIRECTORY code in m_p_i(), which
makes "git diff HEAD submodule/" and "git diff HEAD submodule" produce
the same output. Previously only the version without trailing slash
returns the difference (if any).

That's the effect of new ce_path_match(). dir_path_match() is not
executed by the new tests. And it should not introduce regressions.

Previously if path "dir/" is passed in with pathspec "dir/", they
obviously match. With new dir_path_match(), the path becomes
_directory_ "dir" vs pathspec "dir/", which is not executed by the old
code path in m_p_i(). The new code path is executed and produces the
same result.

The other case is pathspec "dir" and path "dir/" is now turned to
"dir" (with DO_MATCH_DIRECTORY). Still the same result before or after
the patch.

So why change? Because of the next patch about clean.c.

Signed-off-by: Nguyễn Thái Ngọc Duy <pclouds@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 5b5490541108913a9f53b917e922ee8f7fe2fac5 [formerly 4157c18b3667dfabd6855ccd8fb9f8897d73617b] [formerly 4157c18b3667dfabd6855ccd8fb9f8897d73617b [formerly ae8d0824217bdf97c69ead49568cd03fc140627b]]
Former-commit-id: 3f55d585e30ffecff5d20b72d822558627519a5d
Former-commit-id: fc7963d0ec2996da004b50759a48919124eb42bb