Documentation: describe 'submodule update --remote' use case
Make it clear that there is no implicit floating going on; --remote
lets you explicitly integrate the upstream branch in your current
HEAD (just like running 'git pull' in the submodule).  The only
distinction with the current 'git pull' is the config location and
setting used for the upstream branch, which is hopefully clear now.

Signed-off-by: W. Trevor King <wking@tremily.us>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 8d1d07e19ee73d95036f8c3827762cc1c2adace9 [formerly 8d1d07e19ee73d95036f8c3827762cc1c2adace9 [formerly 9937e65d8835ee6c026892de48f993d7a524a00f]]
Former-commit-id: a5d3d4a55f118be15884aea438cdd190e7e25038
Former-commit-id: bd3132aa80ccb41e744161f4c55378229e40a65f