merge-recursive.c: tolerate missing files while refreshing index
Teach add_cacheinfo to tell make_cache_entry to skip refreshing stat
information when a file is missing from the work tree.  We do not want
the index to be stat-dirty after the merge but also do not want to fail
when a file happens to be missing.

This fixes the 'merge-recursive w/ empty work tree - ours has rename'
case in t3030-merge-recursive.

Suggested-by: Elijah Newren <newren@gmail.com>
Signed-off-by: Brad King <brad.king@kitware.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: dc9d860f81e74a4f6efd2ede8a955987272c7e43 [formerly dc9d860f81e74a4f6efd2ede8a955987272c7e43 [formerly 6e2068ae48000a2dfdb2044bbb91073c11f6fbff]]
Former-commit-id: 869a4d97a6b31b6fd44cab6df19224f4d58a09c8
Former-commit-id: 70d752dcbca57e7e48310d7b37924e649c0ef960