expand_user_path: do not look at NULL path
We explicitly check for and handle the case that the
incoming "path" variable is NULL, but before doing so we
call strchrnul on it, leading to a potential segfault.

We can fix this simply by moving the strchrnul call down; as
a bonus, we can tighten the scope on the associated
variable.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 19f2619e83c700efc89090c7fcbe2193d7ff8cca [formerly 30ca689ec65a10c54befbc602c2fd53406b7cd16] [formerly 30ca689ec65a10c54befbc602c2fd53406b7cd16 [formerly 53ec551c87c731c5c4171e943842998bdfb5548e]]
Former-commit-id: 70abe50e93624c21e01bc9d816e900451c2c400e
Former-commit-id: 41fe425078f67f54f4b271ddb904e9fa231f1a8d