request-pull: pick up tag message as before
The previous two steps were meant to stop updating the explicit
refname the user gave to the command to a different ref that points
at it.  Most notably, we no longer substitute a branch name the user
used with a name of the tag that points at the commit at the tip of
the branch (it still can be done with "local-branch:remote-tag").

However, they also lost the code that included the message in a
tag when the user _did_ ask the tag to be pulled.  Resurrect it.

Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 439276cf3a072326a59f0a02319b9b1eb0132ed8 [formerly b6eae09485639296d19dbb8baa2ae702a329cf26] [formerly b6eae09485639296d19dbb8baa2ae702a329cf26 [formerly 4b14ec878a206e87520378b9e1e43aef469bc646]]
Former-commit-id: 34ef873951181b917f231756788158416cdb0f69
Former-commit-id: c4f93cb75dd91bbdaefd989777d594a77060a4dc