reset: support "--mixed --intent-to-add" mode
When --mixed is used, entries could be removed from index if the
target ref does not have them. When "reset" is used in preparation for
commit spliting (in a dirty worktree), it could be hard to track what
files to be added back. The new option --intent-to-add simplifies it
by marking all removed files intent-to-add.

Signed-off-by: Nguyễn Thái Ngọc Duy <pclouds@gmail.com>

Former-commit-id: 66074547d22a6bf4f01e38576cb1e12602dc7e58 [formerly 66074547d22a6bf4f01e38576cb1e12602dc7e58 [formerly b4b313f94ab06fabe78fa91b3f2ebbc7ef08b012]]
Former-commit-id: 19028be92c02b46dc33de503cd5f0b3018c10071
Former-commit-id: 7fad71b4fe98a0ad6288027a4c695b3e1aa7cd62