t3004: add test for ls-files on symlinks via absolute paths
When symlinks in the working tree are manipulated using the absolute
path, git dereferences them, and tries to manipulate the link target
instead.

This causes most high-level functions to misbehave when acting on
symlinks given via absolute paths. For example

  $ git add /dir/repo/symlink

attempts to add the target of the symlink rather than the symlink
itself, which is usually not what the user intends to do.

This is a regression introduced by 18e051a:
  setup: translate symlinks in filename when using absolute paths
(which did not take symlinks inside the work tree into consideration).

Add a known-breakage test using the ls-files function, checking both if
the symlink leads to a target in the same directory, and a target in the
above directory.

Signed-off-by: Martin Erik Werner <martinerikwerner@gmail.com>
Tested-by: Martin Erik Werner <martinerikwerner@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 670b335d0a2d3bff14cdb1ed86c20d34a0c0c57a [formerly 670b335d0a2d3bff14cdb1ed86c20d34a0c0c57a [formerly f02033f1d02a55b32815a4702389e4d492f38bcc]]
Former-commit-id: e9d33524f27f255237e6f019427fef717acba8c8
Former-commit-id: 2ba29de712047d7be7dc3e5795640a2a74e8e681