t0060: add test for prefix_path on symlinks via absolute paths
When symlinks in the working tree are manipulated using the absolute
path, git dereferences them, and tries to manipulate the link target
instead.

This applies to most high-level commands but prefix_path is the common
denominator for all of them.

Add a known-breakage tests using the prefix_path function, which
currently uses real_path, causing the dereference.

Signed-off-by: Martin Erik Werner <martinerikwerner@gmail.com>
Reviewed-by: Duy Nguyen <pclouds@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: bf23576cb53695e955d3f29a7a3ac5d904e41640 [formerly a3704f3dcda7d31190f97f94422ff70ef5795e23] [formerly a3704f3dcda7d31190f97f94422ff70ef5795e23 [formerly 74af95d6aac03e1a1227c78d0d0ea8bbc2fa2b4d]]
Former-commit-id: 035f7fbec27b156c1f12cbc889754051a4fcd0a4
Former-commit-id: 5a11397be62f78ea775726d359115bcf2cef8e5f