t0060: add tests for prefix_path when path begins with work tree
One edge-case that isn't currently checked in the tests is the beginning
of the path matching the work tree, despite the target not actually
being the work tree, for example:

  path = /dir/repoa
  work_tree = /dir/repo

should fail since the path is outside the repo. However, if /dir/repoa
is in fact a symlink that points to /dir/repo, it should instead
succeed.

Add two tests covering these cases, since they might be potential
regression points.

Signed-off-by: Martin Erik Werner <martinerikwerner@gmail.com>
Reviewed-by: Duy Nguyen <pclouds@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 15b0a177d27183b2e2cd7934abfb122bccbde4da [formerly 15b0a177d27183b2e2cd7934abfb122bccbde4da [formerly e131daa4c6baa53aba22e5d8bc1defa83b443bf0]]
Former-commit-id: 4b4efb14290fc0475cebe529f5c0ff1ba15a5e07
Former-commit-id: d780352b3dc1c02ab34797d091c34e2e84bcc6e4