setup: add abspath_part_inside_repo() function
In order to extract the part of an absolute path which lies inside the
repo, it is not possible to directly use real_path, since that would
dereference symlinks both outside and inside the work tree.

Add an abspath_part_inside_repo() function which first checks if the
work tree is already the prefix, then incrementally checks each path
level by temporarily NUL-terminating at each '/' and comparing against
the work tree path. If a match is found, it overwrites the input path
with the remainder past the work tree (which will be the part inside the
work tree).

This function is currently only intended for use in
'prefix_path_gently'.

Signed-off-by: Martin Erik Werner <martinerikwerner@gmail.com>
Reviewed-by: Duy Nguyen <pclouds@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 9477df50862931a028dc804beca2a76b1ee9f648 [formerly 3b64c3d68f05b6d174e5fd30e4192ee0b6e0532e] [formerly 3b64c3d68f05b6d174e5fd30e4192ee0b6e0532e [formerly ddc2a6281595fd24ea01497c496f88c40a59562f]]
Former-commit-id: a88f8eacba9b38e43c48cd8fd91feed85f136cd2
Former-commit-id: 49e59a46feea815d27a06689ed71d69a832c0bf6