t0003: do not chdir the whole test process
Moving to some other directory and letting the remainder of the test
pieces to expect that they start there is a bad practice.  The test
that contains chdir itself may fail (or by mistake skipped via the
GIT_SKIP_TESTS mechanism) in which case the remainder may operate on
files in unexpected places.

Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 09668664262417f6ecc4795f92b328540108dc17 [formerly 6f22d671759b52fe278bbb1bfe7f9302c4bf871c] [formerly 6f22d671759b52fe278bbb1bfe7f9302c4bf871c [formerly c4a7bce1b5d1ad18482a87cb80a6bddabb925f62]]
Former-commit-id: 5604d0a372c8bf3a3a02da590a60df44d0e0bcd8
Former-commit-id: dc4a80208ae480db3500b3c3ffdb9afce1e14f93