check-attr: move to the top of working tree when in non-bare repository
Lasse Makholm noticed that running "git check-attr" from a place
totally unrelated to $GIT_DIR and $GIT_WORK_TREE does not give
expected results.  I think it is because the command does not say it
wants to call setup_work_tree().

We still need to support use cases where only a bare repository is
involved, so unconditionally requiring a working tree would not work
well.  Instead, make a call only in a non-bare repository.

We may want to see if we want to do a similar fix in the opposite
direction to check-ignore.  The command unconditionally requires a
working tree, but it should be usable in a bare repository just like
check-attr attempts to be.

Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 96d68cd8fccc8a980fc29f1e0221e61039cebd2a [formerly 96d68cd8fccc8a980fc29f1e0221e61039cebd2a [formerly cdbf623254fc281e42eb41e700ae785813983960]]
Former-commit-id: 522b06932b64394426051a365749b30ac642e8b8
Former-commit-id: 2b2aa074171c5839c0e302ddf3614ec5399abb56