gc: config option for running --auto in background
`gc --auto` takes time and can block the user temporarily (but not any
less annoyingly). Make it run in background on systems that support
it. The only thing lost with running in background is printouts. But
gc output is not really interesting. You can keep it in foreground by
changing gc.autodetach.

Signed-off-by: Nguyễn Thái Ngọc Duy <pclouds@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 29fc8aff8dc24cec7179ce71ec6a84acc873582e [formerly 7b4f65dcee5b02e4a03dd1b2f8ef8cd12efb8ea5] [formerly 7b4f65dcee5b02e4a03dd1b2f8ef8cd12efb8ea5 [formerly 9f673f9477c63433aa0774c11b9e42979721f4d6]]
Former-commit-id: 128012e12d6403a67f56a297355d2af198ba43f2
Former-commit-id: ab81ade172069cf6478e7a17fb05a1c3f183cf58