tests: auto-set LIB_HTTPD_PORT from test name
We set the default apache port for each of the httpd tests
to the 4-digit test number of the test script. We want these
to remain unique so that the tests do not conflict with each
other when run in parallel.

Instead of doing it manually in each test script, let's just
set it from the test name at run time. This is simpler, and
is one less thing to be updated when test scripts are
renamed (e.g., when being re-rolled or when conflicting
after being merged with another topic).

Incidentally, this fixes a case where t5537 and t5538 used
the same port number (5537), and could conflict with each
other when run in parallel.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: f9096b3a6ba192417ea7f85710686c134aaddbb2 [formerly f9096b3a6ba192417ea7f85710686c134aaddbb2 [formerly 3bb486e4397415e1855e5edbf28d8e50c3d45265]]
Former-commit-id: 1c0ae9aaa0c354990ab2ead95bc3d1308a7733ff
Former-commit-id: 47e6eeec1b6dcf36737f19653ba168cd81ad971c