git-contacts: do not fail parsing of good diffs
If a line in a patch starts with "--- " it will be deemed
malformed unless it also contains the proper diff header
format. This situation can happen with a valid patch if
it has a line starting with "-- " and that line is removed.

This patch just removes the check in git-contacts.

Signed-off-by: Lars Gullik BjÃ¸nnes <larsbj@gullik.org>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 007c7c23056360173d72619bd5d4511bbccdfd97 [formerly 007c7c23056360173d72619bd5d4511bbccdfd97 [formerly 8f72011f1c4a626ed1c1e826661d42274f465e29]]
Former-commit-id: 39f1e97183ca48652aa9c0d6887edced9167d57f
Former-commit-id: ed51a041d224af33c9a0099dc378ae5e8424a6c5