reset: optionally setup worktree and refresh index on --mixed
Refreshing index requires work tree.  So we have two options: always
set up work tree (and refuse to reset if failing to do so), or make
refreshing index optional.

As refreshing index is not the main task, it makes more sense to make
it optional. This allows us to still work in a bare repository to update
what is in the index.

Reported-by: Patrick Palka <patrick@parcs.ath.cx>
Signed-off-by: Nguyễn Thái Ngọc Duy <pclouds@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: cd125eae229299c7e8da8d3e96b64e35e44c72af [formerly 496c3860dece82c8bbf85c7ea4a9231a27de68ad] [formerly 496c3860dece82c8bbf85c7ea4a9231a27de68ad [formerly b7756d41dcc6720e6005b663bcb6f8b2f2a66763]]
Former-commit-id: 55a9185a6a5f43a0432e37ab2e6e8ef4476c8bd0
Former-commit-id: bd0502bb40e15dbed017be02736368b88cb29a0a