diff: do not reuse_worktree_file for submodules
The GIT_EXTERNAL_DIFF calling code attempts to reuse existing worktree
files for the worktree side of diffs, for performance reasons.
However, that code also tries to do the same with submodules.  This
results in calls to $GIT_EXTERNAL_DIFF where the old-file is a file of
the form "Submodule commit $sha1", but the new-file is a directory in
the worktree.

Fix it by never reusing a worktree "file" in the submodule case.

Reported-by: Grégory Pakosz <gregory.pakosz@gmail.com>
Signed-off-by: Thomas Rast <tr@thomasrast.ch>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 24aa5c7802d8fcb706c573ebf0c7d96a635d7bc4 [formerly 24aa5c7802d8fcb706c573ebf0c7d96a635d7bc4 [formerly aba4727281612c3e24914691727e11e1f44a9aac]]
Former-commit-id: b81389f1cfc8687f9a334ba087a2b865a26814b0
Former-commit-id: 04297ee2da26d8824ccb9229aebd24b78ab4e500