gitweb: Avoid overflowing page body frame with large images
When displaying a blob in gitweb, if it's an image, specify constraints for
maximum display width and height to prevent the image from overflowing the
frame of the enclosing page_body div.

This change assumes that it is more desirable to see the whole image without
scrolling (new behavior) than it is to see every pixel without zooming
(previous behavior).

Signed-off-by: Andrew Keller <andrew@kellerfarm.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: c795cad106368ee9d9510f34cdada61e92684fcc [formerly c795cad106368ee9d9510f34cdada61e92684fcc [formerly 46a7471f0e20af9d2efcc4a14c76d537e1551ac9]]
Former-commit-id: c86ea380ad1a10ca035507cdcd60728f2f263a7c
Former-commit-id: 7c694d4ed645be3f311057771ab5c014cda72af1