config: disallow relative include paths from blobs
When we see a relative config include like:

  [include]
  path = foo

we make it relative to the containing directory of the file
that contains the snippet. This makes no sense for config
read from a blob, as it is not on the filesystem.  Something
like "HEAD:some/path" could have a relative path within the
tree, but:

  1. It would not be part of include.path, which explicitly
     refers to the filesystem.

  2. It would need different parsing rules anyway to
     determine that it is a tree path.

The current code just uses the "name" field, which is wrong.
Let's split that into "name" and "path" fields, use the
latter for relative includes, and fill in only the former
for blobs.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Kirill A. Shutemov <kirill@shutemov.name>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 199e863144a778087bce69ff1b22e40c9a138a1d [formerly ca8e00f69d79f32ecdf1f3cd2789605a45ffc66f] [formerly ca8e00f69d79f32ecdf1f3cd2789605a45ffc66f [formerly d14d42440d8370f8fe5016a6e212d101745f70cc]]
Former-commit-id: 64d3be30c694386b08fc1d49bd9f55c43d3ede26
Former-commit-id: 29c13c1fad0e2d8f2f00ddd908fd9fdb647ca981