open_istream(): do not dereference NULL in the error case
When stream-filter cannot be attached, it is expected to return NULL,
and we should close the stream we opened and signal an error by
returning NULL ourselves from this function.

However, we attempted to dereference that NULL pointer between the
point we detected the error and returned from the function.

Brought-to-attention-by: John Keeping <john@keeping.me.uk>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: a991cbf43b65bbec0c1c532d60c4e0fb54fa2094 [formerly a991cbf43b65bbec0c1c532d60c4e0fb54fa2094 [formerly 78368f2c1ad342719ccf1e719bd5126ca6c7b68b]]
Former-commit-id: 8932125d82c4a83913d1d329b8982fa3e7661991
Former-commit-id: e344c54255b7b50aad7e06d490688f49ea79862f