transport-helper.c: do not overwrite forced bit
If the the transport helper says it was a forced update, then it is
a forced update.  It is however possible that an update is forced
without the transport-helper knowing about it, namely because some
higher up code had objections to the update and needed forcing in
order to let it through to the transport helper.  In other words, it
does not necessarily mean the update was *not* forced, when the
helper did not say "forced update".

Signed-off-by: Max Horn <max@quendi.de>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 7f2948eb0dfd3a2a1fb0f1509f7cf25248010fde [formerly 66b38e21e87a0580af3924cf26af336c6e166ca4] [formerly 66b38e21e87a0580af3924cf26af336c6e166ca4 [formerly cf31f70c088a314a9d11a4106d2fe1036051890a]]
Former-commit-id: 663492da40d34fc3a4ac0ef4da1a78f65eacadee
Former-commit-id: efb75eee05f6d26769223edd640012f6d4ec718b