test-lib: allow setting the index format version
Allow adding a TEST_GIT_INDEX_VERSION variable to config.mak to set the
index version with which the test suite should be run.

If it isn't set, the default version given in the source code is
used (currently version 3).

To avoid breakages with index versions other than [23], also set the
index version under which t2104 is run to 3.  This test only tests
functionality specific to version 2 and 3 of the index file and would
fail if the test suite is run with any other version.

Helped-by: Junio C Hamano <gitster@pobox.com>
Signed-off-by: Thomas Gummerer <t.gummerer@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: b72e97a79e38936e7d9eda2747e03fda117d7f14 [formerly b72e97a79e38936e7d9eda2747e03fda117d7f14 [formerly 5d9fc888b485caa1b6a95f2cb583dc26ab7e8d13]]
Former-commit-id: dbdfc2175de255624cbdfaff23ba17d7c94a8683
Former-commit-id: 842c0c93112e37c7b42020c8b4ad2dce0b1da355