difftool: support repositories with .git-files
Modern versions of "git submodule" use .git-files to setup the
submodule directory.  When run in a "git submodule"-created
repository "git difftool --dir-diff" dies with the following
error:

	$ git difftool -d HEAD~
	fatal: This operation must be run in a work tree
	diff --raw --no-abbrev -z HEAD~: command returned error: 128

core.worktree is relative to the .git directory but the logic
in find_worktree() does not account for it.

Use `git rev-parse --show-toplevel` to find the worktree so that
the dir-diff feature works inside a submodule.

Reported-by: Gábor Lipták <gabor.liptak@gmail.com>
Helped-by: Jens Lehmann <jens.lehmann@web.de>
Helped-by: John Keeping <john@keeping.me.uk>
Signed-off-by: David Aguilar <davvid@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 44be05e05c7ea9f0b7a62925ecd963b4c2284566 [formerly 44be05e05c7ea9f0b7a62925ecd963b4c2284566 [formerly 94eaa806519498de2ca59a424b013812d72d21c5]]
Former-commit-id: 58bd4491bd744f8a7c2efd1de5ca854b848feb03
Former-commit-id: ba2caaf4b4bcfbd45d5ad0b91d9add414d73eee4