date: check date overflow against time_t
When we check whether a timestamp has overflowed, we check
only against ULONG_MAX, meaning that strtoul has overflowed.
However, we also feed these timestamps to system functions
like gmtime, which expect a time_t. On many systems, time_t
is actually smaller than "unsigned long" (e.g., because it
is signed), and we would overflow when using these
functions.  We don't know the actual size or signedness of
time_t, but we can easily check for truncation with a simple
assignment.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 7cb107a7232ed3e8672fdd79f916307f3b95a16b [formerly ee63e8d2be68ea33bd2f3b9802f1838f14494e5b] [formerly ee63e8d2be68ea33bd2f3b9802f1838f14494e5b [formerly 7ca36d9398a85e7974d04f8fbd2c6adb088290e1]]
Former-commit-id: 6c6656555a1480d8b1d666b28aabd22c5bf02d48
Former-commit-id: c33675126b03e8751ec88289a6519e32c5ae4c37