remote: handle pushremote config in any order
The remote we push can be defined either by
remote.pushdefault or by branch.*.pushremote for the current
branch. The order in which they appear in the config file
should not matter to precedence (which should be to prefer
the branch-specific config).

The current code parses the config linearly and uses a
single string to store both values, overwriting any
previous value. Thus, config like:

  [branch "master"]
  pushremote = foo
  [remote]
  pushdefault = bar

erroneously ends up pushing to "bar" from the master branch.

We can fix this by storing both values and resolving the
correct value after all config is read.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 6de0468007d862c41ea6a61325feee6b8a53d250 [formerly cc28894aec70399550e28ec6d7120e75e8d02f3e] [formerly cc28894aec70399550e28ec6d7120e75e8d02f3e [formerly 98b406f3ad6a6989a5b11c2a2582a9f539d66263]]
Former-commit-id: 9c73560c24cc0edc7869f2aaacfc18f7cc3eb67a
Former-commit-id: 146a39b0cd3d005b0a199e5824e93af72d53b7bc