tree-diff: move all action-taking code out of compare_tree_entry()
- let it do only comparison.

This way the code is cleaner and more structured - cmp function only
compares, and the driver takes action based on comparison result.

There should be no change in performance, as effectively, we just move
if series from on place into another, and merge it to was-already-there
same switch/if, so the result is maybe a little bit faster.

Signed-off-by: Kirill Smelkov <kirr@mns.spb.ru>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 0f032901ecdd4cca62590ae1643386dea97199d6 [formerly 0f032901ecdd4cca62590ae1643386dea97199d6 [formerly 903bba68abb8153f181d854e9a7075b893f1303f]]
Former-commit-id: 6864aa2ac1d03d659fd01d116d7cf8330181279b
Former-commit-id: 249d03df0450b2b8a25942e9b9b4447fc9c6f075