tree-diff: rename compare_tree_entry -> tree_entry_pathcmp
Since previous commit, this function does not compare entry hashes, and
mode are compared fully outside of it. So what it does is compare entry
names and DIR bit in modes. Reflect this in its name.

Add documentation stating the semantics, and move the note about
files/dirs comparison to it.

Signed-off-by: Kirill Smelkov <kirr@mns.spb.ru>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: d4ab043b13dd9d88689b7c8e072adfacba33df9e [formerly 7be3e5c42b7fdbe0d97b6f22cb325a205b8c7d71] [formerly 7be3e5c42b7fdbe0d97b6f22cb325a205b8c7d71 [formerly 9bc0619655dfd183eb5d0d47b4d63a511acd5c64]]
Former-commit-id: 69e4d665b609e54ae09dc41e78102029c1295ae5
Former-commit-id: c21ce502a2671949218c74f9c585963d7befb511