am doc: add a pointer to relevant hooks
It is not obvious when looking at a new command what hooks will affect
it.  Add a HOOKS section to the git-am(1) page, imitating
git-commit(1), to make it easier for people to discover e.g. the
applypatch-msg hook that can implement a custom subject-mangling
strategy (e.g., removing a "bug #nnnn:" prefix introduced by a bug
tracker).

Reported-by: Phillip Susi <psusi@ubuntu.com>
Signed-off-by: Jonathan Nieder <jrnieder@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 17feb4f5be1f92f9f37432759548bb440e062e11 [formerly 17feb4f5be1f92f9f37432759548bb440e062e11 [formerly 0e3d40c60d9ae14ddc9f525ce4db3cbc0bde5278]]
Former-commit-id: 4a117064426892baac2d146936a020020143fc31
Former-commit-id: aeac27cd29c453bb3f77ab86e4aed939e1b57f97