request-pull: resurrect "pretty refname" feature
When asking to fetch/pull a branch whose name is B or a tag whose
name is T, we used to show the command to run as:

	git pull $URL B
        git pull $URL tags/T

even when B and T were spelled in a more qualified way in order to
disambiguate, e.g. heads/B or refs/tags/T, but the recent update
lost this feature.  Resurrect it.

Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 57f178675a00552111d0bd42385f36c0cd265edd [formerly 57f178675a00552111d0bd42385f36c0cd265edd [formerly 5aae66bd998429959d89c5bedf0bd8175a2b3f33]]
Former-commit-id: 7b9c771881527e57d35143acb30a5e7c7f93bb5b
Former-commit-id: da8fa39c501db0f08790ab7bcdd3c38dc8250fe0