fetch: add a failing test for prunning with overlapping refspecs
When a remote has multiple fetch refspecs and these overlap in the
target namespace, fetch may prune a remote-tracking branch which still
exists in the remote. The test uses a popular form of this, by putting
pull requests as stored in a popular hosting platform alongside "real"
remote-tracking branches.

The fetch command makes a decision of whether to prune based
on the first matching refspec, which in this case is insufficient, as it
covers the pull request names. This pair of refspecs does work as
expected if the more "specific" refspec is the first in the list.

Signed-off-by: Carlos Martín Nieto <cmn@elego.de>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 9f6aad0fa4232bae3900897a7e9e43ca5e309434 [formerly 9f6aad0fa4232bae3900897a7e9e43ca5e309434 [formerly f377e7a37c1b28359a228cf5bb43161a8a22b385]]
Former-commit-id: 91e8be48baa4f6bef96cdf88bc0ec08ffc111df1
Former-commit-id: 0000d631f91299ffe2cc49d86609b8808b3b5ee7