fetch: handle overlaping refspecs on --prune
We need to consider that a remote-tracking branch may match more than
one rhs of a fetch refspec. In such a case, it is not enough to stop at
the first match but look at all of the matches in order to determine
whether a head is stale.

To this goal, introduce a variant of query_refspecs which returns all of
the matching refspecs and loop over those answers to check for
staleness.

Signed-off-by: Carlos Martín Nieto <cmn@elego.de>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: bf33a9319791f7b86dd4da9c28149811a99fb3dc [formerly bf33a9319791f7b86dd4da9c28149811a99fb3dc [formerly e6f637122ecd60c59df27b8188b6ddac39ac0063]]
Former-commit-id: 5946bfbca69aac87cbacbfab12a67148a5a778f0
Former-commit-id: ab36960f8e7695ac39d73e96208fcfda4f872500