shallow: use stat_validity to check for up-to-date file
When we are about to write the shallow file, we check that
it has not changed since we last read it. Instead of
hand-rolling this, we can use stat_validity. This is built
around the index stat-check, so it is more robust than just
checking the mtime, as we do now (it uses the same check as
we do for index files).

The new code also handles the case of a shallow file
appearing unexpectedly. With the current code, two
simultaneous processes making us shallow (e.g., two "git
fetch --depth=1" running at the same time in a non-shallow
repository) can race to overwrite each other.

As a bonus, we also remove a race in determining the stat
information of what we read (we stat and then open, leaving
a race window; instead we should open and then fstat the
descriptor).

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: fc33535bec7eb8843ed19432fb3b2df086009b65 [formerly 9777806e17e890ca1135d7b4a268b085354b66eb] [formerly 9777806e17e890ca1135d7b4a268b085354b66eb [formerly 0cc77c386cea7afebb54a5e7263ca37569ecfe7a]]
Former-commit-id: a021af1e986cf586438477efb818ca71d0244211
Former-commit-id: 3ba93978996264bde0a3dc8f23f207dbb392186a