branch.c: delete size check of newly tracked branch names
Since commit 6f084a56 the length of a newly tracked branch name was limited
to 1019 = 1024 - 7 - 7 - 1 characters, a bound derived by having to store
this name in a char[1024] called key with two strings of length at most 7
and a '\0' character.

This was no longer necessary as of commit a9f2c136, which uses a strbuf
(documented in Documentation/technical/api-strbuf.txt) to store this value.

Remove this unneeded check to allow branch names longer than 1019
characters.

Signed-off-by: Jacopo Notarstefano <jacopo.notarstefano@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: cd77100f95c864e5d55eaf4da492e52e536f2d61 [formerly 79e6b018784bb2fe36f28018ccdf13149de5f8ec] [formerly 79e6b018784bb2fe36f28018ccdf13149de5f8ec [formerly 9ef5e2a7223fbabf6a29946c1c120bc2c3c6cf86]]
Former-commit-id: cedb7fe2b3f2962a7f39fbf7bb9e0f93cbf1526d
Former-commit-id: bb539998cc1e14a3fd8ee13dbf0022781bc3048e