submodule update: consistently document the '--checkout' option
Commit 322bb6e12f (add update 'none' flag to disable update of submodule
by default) added the '--checkout' option to "git submodule update" but
forgot to explicitly document it in synopsis, usage string and man page
(It is only mentioned implicitly in the man page). In 23d25e48 (submodule:
explicit local branch creation in module_clone) the synopsis of the man
page was updated, but the "OPTIONS" section of the man page and the usage
string of the git-submodule script still do not mention the '--checkout'
option.

Fix that by documenting this option in usage string and the "OPTIONS"
section of man page too. While at it group the update-mode options into
a single set in the usage string.

Reported-by: Matthijs Kooijman <matthijs@stdin.nl>
Signed-off-by: Jens Lehmann <Jens.Lehmann@web.de>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 439e137d5fbd423cbad542b9c8340e02648f0079 [formerly 439e137d5fbd423cbad542b9c8340e02648f0079 [formerly 893a9764dc5e6d5d3432a47f595ffffade1edd9c]]
Former-commit-id: 9c1b5e5db6c86a1977e02dc40155865fa4657837
Former-commit-id: b89402800fb211524a5613c7d7065d9a5d9eb05c