match_explicit_lhs: allow a "verify only" mode
The match_explicit_lhs function has all of the logic
necessary to verify the refspecs without actually doing any
work. This patch lets callers pass a NULL "match" pointer to
indicate they want a "verify only" operation.

For the most part, we just need to avoid writing to the NULL
pointer. However, we also have to refactor the
try_explicit_object_name sub-function; it indicates success by
allocating and returning a new ref. Instead, we give it an
"out" parameter for the match and return a numeric status.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 36de05b131bce500d2535fcc222cb0eaeab1a493 [formerly 36de05b131bce500d2535fcc222cb0eaeab1a493 [formerly 471fd3fe410ad95ef11270a865203482ce4bca7d]]
Former-commit-id: 2ae86463f6fe41c194cded5175c2649a45f3d787
Former-commit-id: cdb88a31e0ac91912e7bf5a7a01847264cd733cf