show_ident_date: fix tz range check
Commit 7f6c3e2dab64 [formerly df3a39fb1170] [formerly 065967edf772] [formerly 51b7e3597e7f [formerly 51b7e3597e7f [formerly 1dca155fe3fa]]] (log: handle integer overflow in
timestamps, 2014-02-24) tried to catch integer overflow
coming from strtol() on the timezone field by comparing against
LONG_MIN/LONG_MAX. However, the intermediate "tz" variable
is an "int", which means it can never be LONG_MAX on LP64
systems; we would truncate the output from strtol before the
comparison.

Clang's -Wtautological-constant-out-of-range-compare notices
this and rightly complains.

Let's instead store the result of strtol in a long, and then
compare it against INT_MIN/INT_MAX. This will catch overflow
from strtol, and also overflow when we pass the result as an
int to show_date.

Reported-by: Eric Sunshine <sunshine@sunshineco.com>
Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 934023ea06c9fc7b76566d10b2cdae7a33f7e0c9 [formerly 934023ea06c9fc7b76566d10b2cdae7a33f7e0c9 [formerly 3f419d45ef0dfc33dc301d9ae4737043c091291a]]
Former-commit-id: 1d1449ed85e8f169647070ae9008f222a009e437
Former-commit-id: 0bfeb07fd181580ebc4a4203dbbf7a9a5383ae46